Charts Improvement
