+++ data_Insights_website/static/js/lib/charts/line.js
@@ -1,12 +1,13 @@
 /*
-    #TODO: Implement supports for ecParams.ecParams
-    #TODO: Interpolating polynamials not scaling well, Investigate.
+    #TODO: Implement supports for     #TODO: Interpolating polynamials not scaling well, Investigate.
 */
 var interpolation = require("../charts/interpolation");
 
 
 var line = function(xAxis, yAxis, lines, chartTitle = 'line chart') {
     _lines = [];
+
+    selected = {};
     var magicType = {};
     var formatter;
     var x_type = [{
@@ -26,6 +27,7 @@ var line = function(xAxis, yAxis, lines, chartTitle = 'line chart') {
             var serie = {
                 'name': lines[i],
                 'type': 'line',
+                'smooth': true,
                 'data': yAxis[i],
                 'itemStyle': {
                     normal: {
@@ -47,7 +49,7 @@ var line = function(xAxis, yAxis, lines, chartTitle = 'line chart') {
         }
     }
     if (x_type[0].type === 'value') {
-        for (i = 0; i < lines.length; ++i) {
+        for ( var i = 0; i < lines.length; ++i) {
             var serie = {
                 'name': lines[i],
                 'type': 'line',
@@ -58,6 +60,7 @@ var line = function(xAxis, yAxis, lines, chartTitle = 'line chart') {
                     data: [],
                 }
             };
+            var gy = -Infinity;
             for (var j = 0; j < xAxis.length; ++j) {
                 serie.data.push([xAxis[j], yAxis[i][j]]);
                 var point = {
@@ -65,26 +68,208 @@ var line = function(xAxis, yAxis, lines, chartTitle = 'line chart') {
                     'xAxis': xAxis[j],
                     'yAxis': yAxis[i][j]
                 };
+
+                gy = (yAxis[i][j] > gy ) ? yAxis[i][j]: gy;
                 serie.markPoint.data.push(point);
-            }console.log(serie);
+            }
+            serie.data = serie.data.sort(function(a, b) { return a[0] - b[0]; });
+
+
+
+            selected[lines[i]] = i < 2;
             series.push(serie);
             _lines.push(lines[i]);
-            if (xAxis.length < 25) {
+
+
+            var is_sorted = xAxis.every((val, i, arr) => !i || (val >= arr[i - 1]));
+            var _x = []; var _y = [];
+            if ( is_sorted === false ) {
+
+                for (var z = 0; z < serie.data.length; ++z) {
+                    _x.push(serie.data[z][0]);
+                    _y.push(serie.data[z][1]);
+                }
+            } else {
+                _x = xAxis;
+                _y = yAxis[i];
+            }
+            if ( xAxis.length < 500 ) {
+                data_spline = interpolation.cubic_spline(_x, _y);
+
+                data_spline =  data_spline.sort(function(a, b) { return a[0] - b[0];});
+                console.log(data_spline);
+                spline_serie = {
+                    'name': lines[i] + ' cubic spline ',
+                    'type': 'line',
+                    'smooth': true,
+                    'data': data_spline,
+                };
+                series.push(spline_serie);
+                _lines.push(lines[i] + ' cubic spline ');
+            }
+
+            // Regression 
+            serieIBestLineFit = interpolation.Regressions(xAxis,yAxis[i]);
+            best_fit_line = []
+            var _xAxis = interpolation.gen_xAxis(_x);
+            for ( var k = 0; k < _xAxis.length; ++k) {
+                best_fit_line.push([_xAxis[k], _xAxis[k] * serieIBestLineFit.m + serieIBestLineFit.b])
+            }
+            best_fit_expr = 'Best Fit line\ny = ' + serieIBestLineFit.m.toFixed(3)+'*x + ' + serieIBestLineFit.b.toFixed(3) + '\n' + 'Correlation coefficient: ' + serieIBestLineFit.r.toFixed(3);
+            bestFit = {
+                'name': lines[i] + ' best fit line ',
+                'type': 'line',
+                'smooth': true,
+                'data': best_fit_line,
+                'markPoint': {
+                    itemStyle: {
+                        normal: {
+                            color: 'transparent'
+                        }
+                    },
+                    label: {
+                        normal: {
+                            borderWidth: 1,
+                            borderRadius: 3,
+                            padding: [5, 5, 6, 7],
+                            show: true,
+                            position: 'left',
+                            formatter: best_fit_expr,
+                            textStyle: {
+                                color: '#fff',
+                                backgroundColor: 'rgba(46,46,46,0.9)',
+                                fontSize: 14,
+                                margin: 5
+                            }
+                        }
+                    },
+                    data: [{
+                        coord: [serieIBestLineFit.gx, gy]
+                    }]
+                }
+            };
+
+            best_fit_name = lines[i] + ' best fit line ';
+            series.push(serie);
+            _lines.push(best_fit_name);
+            selected[best_fit_name] = false;
+            series.push(bestFit);
+
+            if (isNaN(serieIBestLineFit.A) !== true &&  isNaN(serieIBestLineFit.R) != true) {
+                best_fit_exp = [];
+                for ( var k = 0; k < _xAxis.length; ++k) {
+                    best_fit_exp.push([_xAxis[k], Math.pow(serieIBestLineFit.R,  _xAxis[k]) * serieIBestLineFit.A])
+                }
+                best_fit_expr = 'Best Fit exponential\ny = ' + serieIBestLineFit.A.toFixed(3)+'* (' + serieIBestLineFit.R.toFixed(3) + '^x)';
+                bestFitExponetial = {
+                    'name': lines[i] + ' best fit exponential',
+                    'type': 'line',
+                    'smooth': true,
+                    'data': best_fit_exp,
+                    'markPoint': {
+                        itemStyle: {
+                            normal: {
+                                color: 'transparent'
+                            }
+                        },
+                        label: {
+                            normal: {
+                                borderWidth: 1,
+                                borderRadius: 3,
+                                padding: [5, 5, 6, 7],
+                                show: true,
+                                position: 'left',
+                                formatter: best_fit_expr,
+                                textStyle: {
+                                    color: '#fff',
+                                    backgroundColor: 'rgba(46,46,46,0.9)',
+                                    fontSize: 14,
+                                    margin: 5
+                                }
+                            }
+                        },
+                        data: [{
+                            coord: [serieIBestLineFit.gx, serieIBestLineFit.A*Math.pow(serieIBestLineFit.R, serieIBestLineFit.gx)]
+                        }]
+                    }
+                } 
+
+                best_fit_exp = lines[i] + ' best fit exponential';
+                _lines.push(best_fit_exp);
+                series.push(bestFitExponetial);
+                selected[best_fit_exp] = false;
+            }
+
+            // best fit quadratic
+            best_fit_quadratic = []
+            for ( var k = 0; k < _xAxis.length; ++k) {
+                var x  = _xAxis[k];
+                best_fit_quadratic.push([x, serieIBestLineFit.quadratic_coeffs[0][0]*Math.pow(x, 2) + serieIBestLineFit.quadratic_coeffs[1][0]*x + serieIBestLineFit.quadratic_coeffs[2][0]] )
+            }
+            best_fit_expr = 'Best Fit quadratic\ny = (' + serieIBestLineFit.quadratic_coeffs[0][0].toFixed(3) + ')*x^2 + (' + serieIBestLineFit.quadratic_coeffs[1][0].toFixed(3)+')*x + (' + serieIBestLineFit.quadratic_coeffs[2][0].toFixed(3) + ')';
+            
+            bestFitQuadratic = {
+                'name': lines[i] + ' best fit quadratic',
+                'type': 'line',
+                'smooth': true,
+                'data': best_fit_quadratic,
+                'markPoint': {
+                    itemStyle: {
+                        normal: {
+                            color: 'transparent'
+                        }
+                    },
+                    label: {
+                        normal: {
+                            borderWidth: 1,
+                            borderRadius: 3,
+                            padding: [5, 5, 6, 7],
+                            show: true,
+                            position: 'left',
+                            formatter: best_fit_expr,
+                            textStyle: {
+                                color: '#fff',
+                                backgroundColor: 'rgba(46,46,46,0.9)',
+                                fontSize: 14,
+                                margin: 5
+                            }
+                        }
+                    },
+                    data: [{
+                        coord: [serieIBestLineFit.gx, serieIBestLineFit.quadratic_coeffs[0][0]*Math.pow(serieIBestLineFit.gx, 2) + serieIBestLineFit.quadratic_coeffs[1][0]*serieIBestLineFit.gx + serieIBestLineFit.quadratic_coeffs[2][0]]
+                    }]
+                }
+            };
+
+            best_fit_name = lines[i] + ' best fit quadratic';
+            series.push(serie);
+            _lines.push(best_fit_name);
+            series.push(bestFitQuadratic);
+            selected[best_fit_name] = false;
+
+            /*
+            if (xAxis.length < 15) {
+                var push_chart = true;
                 interp = interpolation.lagrange(xAxis, yAxis[i]);
                 var interp_serie = {
                     'name': 'lagrange ' + lines[i],
                     'type': 'line',
-                    'data': [],
                 };
-                interp_serie.markPoint = serie.markPoint;
+                var interp_data = [];
                 for (var k = 0; k < interp._xAxis.length; ++k) {
-                    interp_serie.data.push([interp._xAxis[k], interp._yAxis[k]])
+                    if ( isNaN(interp._yAxis[k]) ===  true || Math.abs(interp._yAxis) === +Infinity) {
+                        push_chart = false;
+                    }
+                    interp_data.push([interp._xAxis[k], interp._yAxis[k]])
                 }
-                series.push(interp_serie);
-                _lines.push('lagrange ' + lines[i]);
-
+                if (push_chart === true ) {
+                    interp_serie['data'] = interp_data;
+                    series.push(interp_serie);
+                    _lines.push('lagrange ' + lines[i]);
+                }
 
                 interp_newton = interpolation.newton(xAxis, yAxis[i]);
+                push_chart = true;
                 var interp_newton_serie = {
                     'name': 'newton ' + lines[i],
                     'type': 'line',
@@ -92,11 +277,16 @@ var line = function(xAxis, yAxis, lines, chartTitle = 'line chart') {
                 };
                 interp_newton_serie.markPoint = serie.markPoint;
                 for (var k = 0; k < interp_newton._xAxis.length; ++k) {
+                    if ( isNaN(interp._yAxis[k]) ===  true || Math.abs(interp._yAxis) === +Infinity) {
+                        push_chart = false;
+                    }
                     interp_newton_serie.data.push([interp_newton._xAxis[k], interp_newton._yAxis[k]])
                 }
-                series.push(interp_newton_serie);
-                _lines.push('newton ' + lines[i]);
-            }
+                if ( push_chart === true ) {
+                    series.push(interp_newton_serie);
+                    _lines.push('newton ' + lines[i]);
+                }
+            }*/
             formatter = {
                 formatter: function(params) {
                     result = params.seriesName + ': ' + params.name + '<br>';
@@ -112,16 +302,31 @@ var line = function(xAxis, yAxis, lines, chartTitle = 'line chart') {
         title: {
             text: chartTitle,
         },
-        legend: {
-            data: _lines
+        legend: {      
+            type: 'scroll',
+            right: 150,
+            left: 100,
+            top: 5,
+            bottom: 100,
+            data: _lines,
+            selected: selected
         },
         dataZoom: {
             show: true,
             realtime: true,
             start: 0,
             end: 100
+        },tooltip: {
+            trigger: 'axis',
+            axisPointer: {
+                show: true,
+                type: 'cross',
+                lineStyle: {
+                    type: 'dashed',
+                    width: 1
+                }
+            }
         },
-        tooltip: formatter,
         toolbox: {
             show: true,
             feature: {
