+++ data_Insights_website/static/js/lib/charts/theme/mime.js
@@ -1,8 +1,8 @@
 var theme = {
-    backgroundColor: 'rgba(46,46,46,0.5)',
-    color: ['rgb(197, 52, 64)', '#2098d1', '#66a61e', '#984ea3', '#00d2d5',
-        '#ff7f00', '#af8d00', '#7f80cd', '#b3e900', '#c42e60',
-        '#a65628', '#f781bf', '#8dd3c7', '#bebada', '#fb8072',
+    backgroundColor: 'rgba(36,46,46,0.5)',
+    color: ['#2098d1', 'rgb(102, 166, 30)', '#ff0000', 'rgb(152, 78, 163)', 'rgb(0, 210, 213)',
+        'rgb(255, 127, 0)', 'rgb(157, 141, 0)', 'rgb(127, 128, 205)', 'rgb(179, 233, 0)', 'rgb(255, 0, 41)',
+        'rgb(102, 166, 30)', 'rgb(247, 129, 191)', '#8dd3c7', '#bebada', '#fb8072',
         '#6699FF', '#ff6666', '#3cb371', '#b8860b', '#fdb462'
     ],
     title: {
