+++ data_Insights_website/static/js/lib/math.js
@@ -1,1642 +1,3 @@
-(function(){function r(e,n,t){function o(i,f){if(!n[i]){if(!e[i]){var c="function"==typeof require&&require;if(!f&&c)return c(i,!0);if(u)return u(i,!0);var a=new Error("Cannot find module '"+i+"'");throw a.code="MODULE_NOT_FOUND",a}var p=n[i]={exports:{}};e[i][0].call(p.exports,function(r){var n=e[i][1][r];return o(n||r)},p,p.exports,r,e,n,t)}return n[i].exports}for(var u="function"==typeof require&&require,i=0;i<t.length;i++)o(t[i]);return o}return r})()({1:[function(require,module,exports){
-var bar = function(ecParams) {
-    var series = [];
-    var zoom_size = 0;
-    for (i = 0; i < ecParams.series.length; ++i) {
-        var serie = {
-            'name': ecParams.series[i],
-            'type': 'bar',
-            'data': ecParams.yAxis[i],
-        }
-        series.push(serie);
-        zoom_size += ecParams.yAxis[i].length;
-    }
-    zoom_size = Math.ceil(12000 / zoom_size);
-    if (zoom_size > 100) {
-        zoom_size = 100;
-    }
-    zoom_begin = Math.floor((100 - zoom_size) / 2);
-    var option = {
-        title: {
-            text: ecParams.title,
-        },
-        legend: {
-            data: ecParams.series
-        },
-        dataZoom: {
-            show: true,
-            realtime: true,
-            start: zoom_begin,
-            end: zoom_begin + zoom_size,
-        },
-        tooltip: {
-            show: true,
-        },
-        toolbox: {
-            show: true,
-            feature: {
-                dataZoom: {
-                    show: true
-                },
-                dataView: {
-                    show: true,
-                    readOnly: false
-                },
-
-                magicType: {
-                    show: true,
-                    type: ['line', 'bar']
-                },
-                restore: {
-                    show: true
-                },
-                saveAsImage: {
-                    show: true
-                }
-            }
-        },
-        xAxis: [{
-            'type': 'category',
-            'data': ecParams.xAxis
-        }],
-        yAxis: [{
-            type: 'value',
-            scale: true,
-        }],
-        series: series,
-
-    };
-    return option;
-}
-
-module.exports = {
-    chart: bar,
-}
-
-},{}],2:[function(require,module,exports){
-var ecParams = {
-    title: '',
-    subtitle: '',
-
-    parallel: [],
-    parallel_data: [],
-
-    // these are used for line, scatter and bar charts
-    xAxis: [],
-    yAxis: [],
-    zAxis: [],
-
-    // these are used for pie charts
-    values: [],
-    categories: [],
-
-    // this can be use for any chart
-    _data_generator: [{}],
-    series: [],
-}
-
-module.exports = {
-    ecParams: ecParams
-}
-
-},{}],3:[function(require,module,exports){
-var charts = window.charts || {};
-
-charts.line = require('../charts/line');
-charts.bar = require('../charts/bar');
-charts.scatter = require('../charts/scatter');
-charts.k = require('../charts/k');
-charts.pie= require('../charts/pie');
-charts.scatter3D = require('../charts/scatter3D');
-charts.ecParams = require('../charts/ecParams');
-charts.mime = require('../charts/theme/mime');
-charts.parallel = require('../charts/parallel');
-
-window.charts = charts;
-},{"../charts/bar":1,"../charts/ecParams":2,"../charts/k":5,"../charts/line":6,"../charts/parallel":7,"../charts/pie":8,"../charts/scatter":9,"../charts/scatter3D":10,"../charts/theme/mime":11}],4:[function(require,module,exports){
-/*
- #TODO: Implement a mechanism which ensures that the point are in a strictly increasing
-        Sequence before proceeding to interpolation.
- #TODO: Implement a mechanism which handles x, and y of different size
- #TODO: Complete the cubic spline interpolation 
-*/
-
-var _centralTendency = require ('../stats/central_tendency');
-var mathjs = require('../math.js');
-
-function gen_xAxis(xAxis) {
-    // create a new array _xAxis with more points spaced by 0.1
-    var _xAxis = [];
-    for (var i = 0; i < xAxis.length - 1; ++i) {
-        var xi = xAxis[i];
-        var inc = (xAxis[i+1] - xAxis[i])/15;
-        while (xi < xAxis[i + 1]) {
-            _xAxis.push(xi);
-            xi += inc;
-        }
-    }
-    _xAxis.push(xAxis[xAxis.length - 1]);
-    return _xAxis;
-}
-
-var lagrange = function(xAxis, yAxis) {
-    var _xAxis = gen_xAxis(xAxis);
-    var _yAxis = [];
-    for (var i = 0; i < _xAxis.length; ++i) {
-        var yi = 0;
-        for (var j = 0; j < xAxis.length; ++j) {
-            var expr = yAxis[j];
-            for (var k = 0; k < xAxis.length; ++k) {
-                if (j != k) {
-                    expr = expr * (_xAxis[i] - xAxis[k]) / (xAxis[j] - xAxis[k]);
-                }
-            }
-            yi += expr;
-        }
-        _yAxis.push(yi);
-    }
-    return {
-        _xAxis,
-        _yAxis
-    };
-}
-var newton = function(xAxis, yAxis) {
-    var _xAxis = gen_xAxis(xAxis);
-    var _yAxis = [];
-    var f_diff = [];
-    for (var i = 0; i < xAxis.length; i++) {
-        f_diff[i] = new Array(xAxis.length);
-    }
-    for (var i = 0; i < xAxis.length; ++i) {
-        f_diff[i][0] = yAxis[i];
-    }
-    for (var i = 1; i < xAxis.length; ++i) {
-        for (var j = i; j < xAxis.length; ++j) {
-            f_diff[j][i] = (f_diff[j][i - 1] - f_diff[j - 1][i - 1]) / (xAxis[j] - xAxis[j - i]);
-        }
-    }
-    for (var i = 0; i < _xAxis.length; ++i) {
-        result = 0;
-        term = 1;
-        for (j = 0; j < xAxis.length; ++j) {
-            result += term * f_diff[j][j];
-            term *= (_xAxis[i] - xAxis[j]);
-        }
-        _yAxis[i] = result;
-    }
-    return {
-        _xAxis,
-        _yAxis
-    }
-};
-
-var Regressions = function (xAxis, yAxis ){
-    var yAxislog = [];
-    for ( var i = 0; i < yAxis.length; ++i ) {
-        // yAxislog is used for exponential regression
-        yAxislog.push(Math.log10(yAxis[i]));
-    }
-    // We compute the central tendencies of the series
-    var xCentralTendency = _centralTendency.Mean(xAxis);
-    var yCentralTendency = _centralTendency.Mean(yAxis);
-    var yLogCentralTendency = _centralTendency.Mean(yAxislog);
-    
-    var gx = -Infinity; // greatest x value
-    var lx = +Infinity; // greatest y value
-
-
-    var m = 0; 
-    var n = 0;   // (xi - Mean(x))*(yi - Mean(y))
-    var p = 0;   // (xi - Mean(x))^2
-    var z =0;    // (yi - Mean(x))^2
-    var nl = 0;
-    var xAxis2 = 0; var xAxis3 = 0; var xAxis4 = 0; var yAxis2 = 0;
-    var xy = 0; var x2y = 0; var xAxis1 = 0; var yAxis1 = 0;
-
-    for ( var i = 0; i < xAxis.length; ++i ) {
-        gx = (xAxis[i] > gx ) ? xAxis[i]: gx;
-        lx = (xAxis[i] < lx ) ? xAxis[i]: lx;
-
-        // required to determine best fit line
-        n += xCentralTendency.meanDiff[i]*yCentralTendency.meanDiff[i];
-        p += Math.pow(xCentralTendency.meanDiff[i], 2);
-        z += Math.pow(yCentralTendency.meanDiff[i], 2);
-        
-        // required to determine exponential regressions
-        nl += xCentralTendency.meanDiff[i]*yLogCentralTendency.meanDiff[i];
-
-        // required to determine 
-        xAxis1 += xAxis[i];
-        xAxis2 += Math.pow(xAxis[i], 2);
-        xAxis3 += Math.pow(xAxis[i], 3);
-        xAxis4 += Math.pow(xAxis[i], 4);
-
-        yAxis1 += yAxis[i];
-        yAxis2 += Math.pow(yAxis[i], 2);
-
-        xy += yAxis[i]*xAxis[i];
-        x2y += yAxis[i]*Math.pow(xAxis[i], 2);
-    }
-
-    // best fit line
-    m = n/p;
-    r = n/Math.pow(p*z, 1/2);
-    b = yCentralTendency.mean - m*xCentralTendency.mean;
-
-    // best fit exponential
-    ml = nl/p;
-    bl = yLogCentralTendency.mean - ml*xCentralTendency.mean;
-
-
-    // best fit quadratic
-    matrixA = [[xAxis4, xAxis3, xAxis2],
-               [xAxis3, xAxis2, xAxis1],
-               [xAxis2, xAxis1, xAxis.length]]
-    matrixb = [x2y, xy, yAxis1];
-
-
-    quadratic_coeffs = mathjs.lusolve(mathjs.lup(matrixA), matrixb);
-
-
-    return {
-        'm': m,
-        'b': b,
-        'gx': gx,
-        'lx': lx,
-        'r': r,
-        'A': Math.pow(10, bl),
-        'R': Math.pow(10, ml),
-        'quadratic_coeffs' : quadratic_coeffs._data,
-    }
-}
-function gen_df(xs, ys) {
-    var n = xs.length - 1;  
-    var b = []
-    // Create an n x n matrix A
-    var A = [];
-    for ( var i = 0; i < xs.length; ++i ) {
-        var Ai = [];
-        for ( var j = 0; j < xs.length; ++j) {
-            Ai.push(0);
-        }
-        A.push(Ai);
-        b.push(0);
-    }
-
-    for ( var i = 1; i < n; ++i ) {
-        A[i][i-1] = 1/(xs[i] - xs[i-1]);
-        A[i][i] = 2 * (1/(xs[i] - xs[i-1]) + 1/(xs[i+1] - xs[i])) ;
-        A[i][i+1] = 1/(xs[i+1] - xs[i]);
-        b[i] = 3* ( (ys[i]-ys[i-1])/ ((xs[i] - xs[i-1])*(xs[i] - xs[i-1])) 
-                    +  (ys[i+1]-ys[i])/ ((xs[i+1] - xs[i])*(xs[i+1] - xs[i])) );
-    }
-
-    A[0][0  ] = 2/(xs[1] - xs[0]);
-    A[0][1  ] = 1/(xs[1] - xs[0]);
-    b[0] = 3 * (ys[1] - ys[0]) / ((xs[1]-xs[0])*(xs[1]-xs[0]));
-    
-    A[n][n-1] = 1/(xs[n] - xs[n-1]);
-    A[n][n  ] = 2/(xs[n] - xs[n-1]);
-    b[n] = 3 * (ys[n] - ys[n-1]) / ((xs[n]-xs[n-1])*(xs[n]-xs[n-1]));
-
-    ks = mathjs.lusolve(mathjs.lup(A), b)._data;
-    return ks;
-}
-
-spline = function (x, xs, ys, ks) {
-    var i = 1;
-    while(xs[i]<x) i++;
-        
-    var t = (x - xs[i-1]) / (xs[i] - xs[i-1]);
-     
-    var a =  ks[i-1]*(xs[i]-xs[i-1]) - (ys[i]-ys[i-1]);
-    var b = -ks[i]*(xs[i]-xs[i-1]) + (ys[i]-ys[i-1]);
-        
-    var q = (1-t)*ys[i-1] + t*ys[i] + t*(1-t)*(a*(1-t)+b*t);
-    return q;
-}
-
-
-cubic_spline = function(xs, ys) {
-    var ks = gen_df(xs, ys);
-    var smooth_xs;
-    if ( xs.length < 130 ) {
-        smooth_xs = gen_xAxis(xs);
-    } else {
-        smooth_xs = xs;
-    }
-    console.log(smooth_xs);
-    var serie = [];
-    for ( var i = 0; i < smooth_xs.length; ++i  ) {
-        serie.push([smooth_xs[i], spline(smooth_xs[i], xs, ys, ks)]);
-    }
-
-    return serie;
-}
-
-
-module.exports = {
-    gen_xAxis: gen_xAxis,
-    newton: newton,
-    lagrange: lagrange,
-    cubic_spline: cubic_spline,
-    Regressions: Regressions,
-};
-
-},{"../math.js":12,"../stats/central_tendency":13}],5:[function(require,module,exports){
-/*
-    #TODO: Implement supports for ecParams.ecParams
-*/
-var k = function(xAxis, c_OHLC, title) {
-    var option = {
-        title: {
-            text: title
-        },
-        tooltip: {
-            trigger: 'axis',
-            formatter: function(params) {
-                result = params[0].seriesName + '<br>' + params[0].name + '<br>';
-                result += ('Open: ' + params[0].data[0] + '<br>' +
-                    'High: ' + params[0].data[1] + '<br>' +
-                    'Low: ' + params[0].data[2] + '<br>' +
-                    'Close: ' + params[0].data[3]);
-                return result;
-            }
-        },
-        legend: {
-            data: [title]
-        },
-        toolbox: {
-            show: true,
-            feature: {
-                mark: {
-                    show: true
-                },
-                dataZoom: {
-                    show: true
-                },
-                restore: {
-                    show: true
-                },
-                saveAsImage: {
-                    show: true
-                }
-            }
-        },
-
-        dataZoom: {
-            show: true,
-            realtime: true,
-            start: 0,
-            end: 100
-        },
-        xAxis: [{
-            type: 'category',
-            data: xAxis
-        }],
-        yAxis: [{
-            type: 'value',
-            scale: true,
-            boundaryGap: [0.05, 0.05],
-        }],
-        series: [{
-            name: title,
-            type: 'k',
-            data: c_OHLC,
-        }]
-    };
-    return option;
-}
-
-module.exports = {
-    chart: k,
-}
-},{}],6:[function(require,module,exports){
-/*
-    #TODO: Implement supports for     #TODO: Interpolating polynamials not scaling well, Investigate.
-*/
-var interpolation = require("../charts/interpolation");
-
-
-var line = function(xAxis, yAxis, lines, chartTitle = 'line chart') {
-    _lines = [];
-
-    selected = {};
-    var magicType = {};
-    var formatter;
-    var x_type = [{
-        'type': 'category',
-        'data': xAxis
-    }];
-    if (typeof xAxis[0] === "number") {
-        x_type = [{
-            'type': 'value',
-            scale: true,
-        }];
-    }
-    var series = [];
-    var i = 0;
-    if (x_type[0].type === 'category') {
-        for (i = 0; i < lines.length; ++i) {
-            var serie = {
-                'name': lines[i],
-                'type': 'line',
-                'smooth': true,
-                'data': yAxis[i],
-                'itemStyle': {
-                    normal: {
-                        areaStyle: {
-                            type: 'default'
-                        }
-                    }
-                },
-            }
-            series.push(serie);
-        }
-        _lines = lines;
-        formatter = {
-            trigger: 'axis'
-        };
-        magicType = {
-            show: true,
-            type: ['line', 'bar']
-        }
-    }
-    if (x_type[0].type === 'value') {
-        for ( var i = 0; i < lines.length; ++i) {
-            var serie = {
-                'name': lines[i],
-                'type': 'line',
-                'data': [],
-                'markPoint': {
-                    symbol: 'emptyCircle',
-                    symbolSize: 3,
-                    data: [],
-                }
-            };
-            var gy = -Infinity;
-            for (var j = 0; j < xAxis.length; ++j) {
-                serie.data.push([xAxis[j], yAxis[i][j]]);
-                var point = {
-                    'name': 'p' + j,
-                    'xAxis': xAxis[j],
-                    'yAxis': yAxis[i][j]
-                };
-
-                gy = (yAxis[i][j] > gy ) ? yAxis[i][j]: gy;
-                serie.markPoint.data.push(point);
-            }
-            serie.data = serie.data.sort(function(a, b) { return a[0] - b[0]; });
-
-
-
-            selected[lines[i]] = i < 2;
-            series.push(serie);
-            _lines.push(lines[i]);
-
-
-            var is_sorted = xAxis.every((val, i, arr) => !i || (val >= arr[i - 1]));
-            var _x = []; var _y = [];
-            if ( is_sorted === false ) {
-
-                for (var z = 0; z < serie.data.length; ++z) {
-                    _x.push(serie.data[z][0]);
-                    _y.push(serie.data[z][1]);
-                }
-            } else {
-                _x = xAxis;
-                _y = yAxis[i];
-            }
-            if ( xAxis.length < 30000 ) {
-                data_spline = interpolation.cubic_spline(_x, _y);
-
-                data_spline =  data_spline.sort(function(a, b) { return a[0] - b[0];});
-                console.log(data_spline);
-                spline_serie = {
-                    'name': lines[i] + ' cubic spline ',
-                    'type': 'line',
-                    'smooth': true,
-                    'data': data_spline,
-                };
-                series.push(spline_serie);
-                _lines.push(lines[i] + ' cubic spline ');
-            }
-
-            // Regression 
-            serieIBestLineFit = interpolation.Regressions(xAxis,yAxis[i]);
-            best_fit_line = []
-            var _xAxis = interpolation.gen_xAxis(_x);
-            for ( var k = 0; k < _xAxis.length; ++k) {
-                best_fit_line.push([_xAxis[k], _xAxis[k] * serieIBestLineFit.m + serieIBestLineFit.b])
-            }
-            best_fit_expr = 'Best Fit line\ny = ' + serieIBestLineFit.m.toFixed(3)+'*x + ' + serieIBestLineFit.b.toFixed(3) + '\n' + 'Correlation coefficient: ' + serieIBestLineFit.r.toFixed(3);
-            bestFit = {
-                'name': lines[i] + ' best fit line ',
-                'type': 'line',
-                'smooth': true,
-                'data': best_fit_line,
-                'markPoint': {
-                    itemStyle: {
-                        normal: {
-                            color: 'transparent'
-                        }
-                    },
-                    label: {
-                        normal: {
-                            borderWidth: 1,
-                            borderRadius: 3,
-                            padding: [5, 5, 6, 7],
-                            show: true,
-                            position: 'left',
-                            formatter: best_fit_expr,
-                            textStyle: {
-                                color: '#fff',
-                                backgroundColor: 'rgba(46,46,46,0.9)',
-                                fontSize: 14,
-                                margin: 5
-                            }
-                        }
-                    },
-                    data: [{
-                        coord: [serieIBestLineFit.gx, gy]
-                    }]
-                }
-            };
-
-            best_fit_name = lines[i] + ' best fit line ';
-            series.push(serie);
-            _lines.push(best_fit_name);
-            selected[best_fit_name] = false;
-            series.push(bestFit);
-
-            if (isNaN(serieIBestLineFit.A) !== true &&  isNaN(serieIBestLineFit.R) != true) {
-                best_fit_exp = [];
-                for ( var k = 0; k < _xAxis.length; ++k) {
-                    best_fit_exp.push([_xAxis[k], Math.pow(serieIBestLineFit.R,  _xAxis[k]) * serieIBestLineFit.A])
-                }
-                best_fit_expr = 'Best Fit exponential\ny = ' + serieIBestLineFit.A.toFixed(3)+'* (' + serieIBestLineFit.R.toFixed(3) + '^x)';
-                bestFitExponetial = {
-                    'name': lines[i] + ' best fit exponential',
-                    'type': 'line',
-                    'smooth': true,
-                    'data': best_fit_exp,
-                    'markPoint': {
-                        itemStyle: {
-                            normal: {
-                                color: 'transparent'
-                            }
-                        },
-                        label: {
-                            normal: {
-                                borderWidth: 1,
-                                borderRadius: 3,
-                                padding: [5, 5, 6, 7],
-                                show: true,
-                                position: 'left',
-                                formatter: best_fit_expr,
-                                textStyle: {
-                                    color: '#fff',
-                                    backgroundColor: 'rgba(46,46,46,0.9)',
-                                    fontSize: 14,
-                                    margin: 5
-                                }
-                            }
-                        },
-                        data: [{
-                            coord: [serieIBestLineFit.gx, serieIBestLineFit.A*Math.pow(serieIBestLineFit.R, serieIBestLineFit.gx)]
-                        }]
-                    }
-                } 
-
-                best_fit_exp = lines[i] + ' best fit exponential';
-                _lines.push(best_fit_exp);
-                series.push(bestFitExponetial);
-                selected[best_fit_exp] = false;
-            }
-
-            // best fit quadratic
-            best_fit_quadratic = []
-            for ( var k = 0; k < _xAxis.length; ++k) {
-                var x  = _xAxis[k];
-                best_fit_quadratic.push([x, serieIBestLineFit.quadratic_coeffs[0][0]*Math.pow(x, 2) + serieIBestLineFit.quadratic_coeffs[1][0]*x + serieIBestLineFit.quadratic_coeffs[2][0]] )
-            }
-            best_fit_expr = 'Best Fit quadratic\ny = (' + serieIBestLineFit.quadratic_coeffs[0][0].toFixed(3) + ')*x^2 + (' + serieIBestLineFit.quadratic_coeffs[1][0].toFixed(3)+')*x + (' + serieIBestLineFit.quadratic_coeffs[2][0].toFixed(3) + ')';
-            
-            bestFitQuadratic = {
-                'name': lines[i] + ' best fit quadratic',
-                'type': 'line',
-                'smooth': true,
-                'data': best_fit_quadratic,
-                'markPoint': {
-                    itemStyle: {
-                        normal: {
-                            color: 'transparent'
-                        }
-                    },
-                    label: {
-                        normal: {
-                            borderWidth: 1,
-                            borderRadius: 3,
-                            padding: [5, 5, 6, 7],
-                            show: true,
-                            position: 'left',
-                            formatter: best_fit_expr,
-                            textStyle: {
-                                color: '#fff',
-                                backgroundColor: 'rgba(46,46,46,0.9)',
-                                fontSize: 14,
-                                margin: 5
-                            }
-                        }
-                    },
-                    data: [{
-                        coord: [serieIBestLineFit.gx, serieIBestLineFit.quadratic_coeffs[0][0]*Math.pow(serieIBestLineFit.gx, 2) + serieIBestLineFit.quadratic_coeffs[1][0]*serieIBestLineFit.gx + serieIBestLineFit.quadratic_coeffs[2][0]]
-                    }]
-                }
-            };
-
-            best_fit_name = lines[i] + ' best fit quadratic';
-            series.push(serie);
-            _lines.push(best_fit_name);
-            series.push(bestFitQuadratic);
-            selected[best_fit_name] = false;
-
-            /*
-            if (xAxis.length < 15) {
-                var push_chart = true;
-                interp = interpolation.lagrange(xAxis, yAxis[i]);
-                var interp_serie = {
-                    'name': 'lagrange ' + lines[i],
-                    'type': 'line',
-                };
-                var interp_data = [];
-                for (var k = 0; k < interp._xAxis.length; ++k) {
-                    if ( isNaN(interp._yAxis[k]) ===  true || Math.abs(interp._yAxis) === +Infinity) {
-                        push_chart = false;
-                    }
-                    interp_data.push([interp._xAxis[k], interp._yAxis[k]])
-                }
-                if (push_chart === true ) {
-                    interp_serie['data'] = interp_data;
-                    series.push(interp_serie);
-                    _lines.push('lagrange ' + lines[i]);
-                }
-
-                interp_newton = interpolation.newton(xAxis, yAxis[i]);
-                push_chart = true;
-                var interp_newton_serie = {
-                    'name': 'newton ' + lines[i],
-                    'type': 'line',
-                    'data': [],
-                };
-                interp_newton_serie.markPoint = serie.markPoint;
-                for (var k = 0; k < interp_newton._xAxis.length; ++k) {
-                    if ( isNaN(interp._yAxis[k]) ===  true || Math.abs(interp._yAxis) === +Infinity) {
-                        push_chart = false;
-                    }
-                    interp_newton_serie.data.push([interp_newton._xAxis[k], interp_newton._yAxis[k]])
-                }
-                if ( push_chart === true ) {
-                    series.push(interp_newton_serie);
-                    _lines.push('newton ' + lines[i]);
-                }
-            }*/
-            formatter = {
-                formatter: function(params) {
-                    result = params.seriesName + ': ' + params.name + '<br>';
-                    result += ('  xAxis :' + params.data.xAxis + '<br>' +
-                        '  yAxis :' + params.data.yAxis);
-                    return result;
-                }
-            }
-
-        }
-    }
-    var options = {
-        title: {
-            text: chartTitle,
-        },
-        legend: {      
-            type: 'scroll',
-            right: 150,
-            left: 100,
-            top: 5,
-            bottom: 100,
-            data: _lines,
-            selected: selected
-        },
-        dataZoom: {
-            show: true,
-            realtime: true,
-            start: 0,
-            end: 100
-        },tooltip: {
-            trigger: 'axis',
-            axisPointer: {
-                show: true,
-                type: 'cross',
-                lineStyle: {
-                    type: 'dashed',
-                    width: 1
-                }
-            }
-        },
-        toolbox: {
-            show: true,
-            feature: {
-                dataZoom: {
-                    show: true
-                },
-                dataView: {
-                    show: true,
-                    readOnly: false
-                },
-                magicType: magicType,
-                restore: {
-                    show: true
-                },
-                saveAsImage: {
-                    show: true
-                }
-            }
-        },
-        xAxis: x_type,
-        yAxis: [{
-            type: 'value',
-            scale: true,
-        }],
-        series: series,
-    }
-
-    return options;
-}
-
-module.exports = {
-    chart: line,
-}
-},{"../charts/interpolation":4}],7:[function(require,module,exports){
-var parallel = function(ecParams) {
-    var data = [];
-    var axis = [];
-    console.log(ecParams);
-    console.log((ecParams.parallel_data[0])[0]);
-    for (var i = 0; i < ecParams.parallel_data[0].length; ++i) {
-        var _instance = [];
-        for (var j = 0; j < ecParams.parallel.length; ++j) {
-            _instance.push(ecParams.parallel_data[j][i]);
-        }
-        data.push(_instance);
-    }
-    for( var i = 0; i < ecParams.parallel.length; ++i ){
-        axis.push({
-            dim: i,
-            name: ecParams.parallel[i],
-        });
-    }
-    var option = {
-        legend: {
-            data: [ecParams.series]
-        },
-        title: {
-            text: ecParams.title,
-        },
-        parallelAxis: axis,
-        series: [{
-            name: ecParams.series,
-            type: 'parallel',
-            data: data,
-        }]
-    };
-
-    return option;
-}
-
-module.exports = {
-    'chart': parallel,
-}
-
-},{}],8:[function(require,module,exports){
-var pie = function(ecParams) {
-    series = [];
-    selected = [];
-    for (var i = 0; i < ecParams.categories.length; ++i) {
-        series.push({
-            value: ecParams.values[i],
-            name: ecParams.categories[i]
-        });
-        selected[ecParams.categories] = i < 10;
-    }
-    var option = {
-        title: {
-            text: ecParams.title,
-        },
-        selected: selected,
-        tooltip: {
-            trigger: 'item',
-            formatter: "{b} : {c} ({d}%)"
-        },
-        legend: {
-            type: 'scroll',
-            orient: 'vertical',
-            right: 20,
-            top: 30,
-            bottom: 20,
-            data: ecParams.categories,
-        },
-        series: [{
-            type: 'pie',
-            data: series,
-        }]
-
-    };
-    return option;
-}
-
-
-module.exports = {
-    chart: pie,
-}
-
-},{}],9:[function(require,module,exports){
-var interpolation = require("./interpolation");
-
-
-var scatter = function(ecParams) {
-    series = [];
-    _series = [];
-    selected = {};
-    for (var i = 0; i < ecParams.series.length; ++i) {
-        _series.push(ecParams.series[i]);
-        serie = {
-            'name': ecParams.series[i],
-            'type': 'scatter',
-            'data': [],
-        }
-        selected[ecParams.series[i]] = i < 1;
-        if (typeof ecParams._data_generator[i] !== 'function') {
-            var gy = -Infinity;
-            for (var j = 0; j < ecParams.xAxis[i].length; ++j) {
-                serie.data.push([ecParams.xAxis[i][j], ecParams.yAxis[i][j]]);
-                gy = (ecParams.yAxis[i][j] > gy ) ? ecParams.yAxis[i][j]: gy;
-            }
-            if (ecParams.xAxis[i].length < 100) {
-                serie.symbolSize = 10;
-            }
-
-            serieIBestLineFit = interpolation.Regressions(ecParams.xAxis[i], ecParams.yAxis[i]);
-            best_fit_line = []
-            var is_sorted = ecParams.xAxis[i].every((val, i, arr) => !i || (val >= arr[i - 1]));
-            var _x = []; var _y = [];
-            if ( is_sorted === false ) {
-                sorted_data = serie.data.sort(function(a, b) { return a[0] - b[0]});
-                for (var z = 0; z < sorted_data.length; ++z) {
-                    _x.push(serie.data[z][0]);
-                    _y.push(serie.data[z][1]);
-                }
-            } else {
-                _x = ecParams.xAxis[i];
-                _y = ecParams.yAxis[i];
-            }
-            var _xAxis = interpolation.gen_xAxis(_x);
-            for ( var k = 0; k < _xAxis.length; ++k) {
-                best_fit_line.push([_xAxis[k], _xAxis[k] * serieIBestLineFit.m + serieIBestLineFit.b])
-            }
-            best_fit_expr = 'Best Fit line\ny = ' + serieIBestLineFit.m.toFixed(3)+'*x + ' + serieIBestLineFit.b.toFixed(3) + '\n' + 'Correlation coefficient: ' + serieIBestLineFit.r.toFixed(3);
-            
-            best_fit_line.push([serieIBestLineFit.gx + 0.5, serieIBestLineFit.m*(serieIBestLineFit.gx +0.5) + serieIBestLineFit.b]);
-            bestFit = {
-                'name': ecParams.series[i] + ' best fit line ',
-                'type': 'line',
-                'data': best_fit_line,
-                'markPoint': {
-                    itemStyle: {
-                        normal: {
-                            color: 'transparent'
-                        }
-                    },
-                    label: {
-                        normal: {
-                            borderWidth: 1,
-                            borderRadius: 3,
-                            padding: [5, 5, 6, 7],
-                            show: true,
-                            position: 'left',
-                            formatter: best_fit_expr,
-                            textStyle: {
-                                color: '#fff',
-                                backgroundColor: 'rgba(46,46,46,0.9)',
-                                fontSize: 14,
-                                margin: 5
-                            }
-                        }
-                    },
-                    data: [{
-                        coord: [serieIBestLineFit.gx, gy]
-                    }]
-                }
-            };
-
-            best_fit_name = ecParams.series[i] + ' best fit line ';
-            series.push(serie);
-            _series.push(best_fit_name);
-            selected[best_fit_name] = i < 1;
-            series.push(bestFit);
-
-            if (isNaN(serieIBestLineFit.A) !== true &&  isNaN(serieIBestLineFit.R) != true) {
-                best_fit_exp = [];
-                for ( var k = 0; k < _xAxis.length; ++k) {
-                    best_fit_exp.push([_xAxis[k], Math.pow(serieIBestLineFit.R,  _xAxis[k]) * serieIBestLineFit.A])
-                }
-                best_fit_expr = 'Best Fit exponential\ny = ' + serieIBestLineFit.A.toFixed(3)+'* (' + serieIBestLineFit.R.toFixed(3) + '^x)';
-            
-                best_fit_exp.push([serieIBestLineFit.gx + 0.5, serieIBestLineFit.A*Math.pow(serieIBestLineFit.R, (serieIBestLineFit.gx +0.5) )]);
-
-                bestFitExponetial = {
-                    'name': ecParams.series[i] + ' best fit exponential',
-                    'type': 'line',
-                    'data': best_fit_exp,
-                    'markPoint': {
-                        itemStyle: {
-                            normal: {
-                                color: 'transparent'
-                            }
-                        },
-                        label: {
-                            normal: {
-                                borderWidth: 1,
-                                borderRadius: 3,
-                                padding: [5, 5, 6, 7],
-                                show: true,
-                                position: 'left',
-                                formatter: best_fit_expr,
-                                textStyle: {
-                                    color: '#fff',
-                                    backgroundColor: 'rgba(46,46,46,0.9)',
-                                    fontSize: 14,
-                                    margin: 5
-                                }
-                            }
-                        },
-                        data: [{
-                            coord: [serieIBestLineFit.gx, serieIBestLineFit.A*Math.pow(serieIBestLineFit.R, serieIBestLineFit.gx)]
-                        }]
-                    }
-                } 
-
-                best_fit_exp = ecParams.series[i] + ' best fit exponential';
-                _series.push(best_fit_exp);
-                series.push(bestFitExponetial);
-                selected[best_fit_exp] = false;
-            }
-
-            // best fit quadratic
-            best_fit_quadratic = []
-            for ( var k = 0; k < _xAxis.length; ++k) {
-                var x  = _xAxis[k];
-                best_fit_quadratic.push([x, serieIBestLineFit.quadratic_coeffs[0][0]*Math.pow(x, 2) + serieIBestLineFit.quadratic_coeffs[1][0]*x + serieIBestLineFit.quadratic_coeffs[2][0]] )
-            }
-            best_fit_quadratic.push([serieIBestLineFit.gx + 0.5, serieIBestLineFit.quadratic_coeffs[0][0]*Math.pow((serieIBestLineFit.gx + 0.5), 2) + serieIBestLineFit.quadratic_coeffs[1][0]*(serieIBestLineFit.gx + 0.5) + serieIBestLineFit.quadratic_coeffs[2][0]]);
-
-            best_fit_expr = 'Best Fit quadratic\ny = (' + serieIBestLineFit.quadratic_coeffs[0][0].toFixed(3) + ')*x^2 + (' + serieIBestLineFit.quadratic_coeffs[1][0].toFixed(3)+')*x + (' + serieIBestLineFit.quadratic_coeffs[2][0].toFixed(3) + ')';
-            
-            bestFitQuadratic = {
-                'name': ecParams.series[i] + ' best fit quadratic',
-                'type': 'line',
-                'data': best_fit_quadratic,
-                'markPoint': {
-                    itemStyle: {
-                        normal: {
-                            color: 'transparent'
-                        }
-                    },
-                    label: {
-                        normal: {
-                            borderWidth: 1,
-                            borderRadius: 3,
-                            padding: [5, 5, 6, 7],
-                            show: true,
-                            position: 'left',
-                            formatter: best_fit_expr,
-                            textStyle: {
-                                color: '#fff',
-                                backgroundColor: 'rgba(46,46,46,0.9)',
-                                fontSize: 14,
-                                margin: 5
-                            }
-                        }
-                    },
-                    data: [{
-                        coord: [serieIBestLineFit.gx, serieIBestLineFit.quadratic_coeffs[0][0]*Math.pow(serieIBestLineFit.gx, 2) + serieIBestLineFit.quadratic_coeffs[1][0]*serieIBestLineFit.gx + serieIBestLineFit.quadratic_coeffs[2][0]]
-                    }]
-                }
-            };
-
-            best_fit_name = ecParams.series[i] + ' best fit quadratic';
-            series.push(serie);
-            _series.push(best_fit_name);
-            series.push(bestFitQuadratic);
-            selected[best_fit_name] = false;
-
-        } else {
-            serie.data = ecParams._data_generator[i]();
-            series.push(serie);
-        }
-        console.log(serie);
-    }
-    var option = {
-        title: {
-            text: ecParams.title,
-        },
-        tooltip: {
-            trigger: 'axis',
-            axisPointer: {
-                show: true,
-                type: 'cross',
-                lineStyle: {
-                    type: 'dashed',
-                    width: 1
-                }
-            }
-        },
-        legend: {
-            type: 'scroll',
-            right: 150,
-            left: 100,
-            top: 5,
-            bottom: 100,
-            data: _series,
-            selected: selected,
-        },
-        dataZoom: {
-            show: true,
-            realtime: true,
-            start: 0,
-            end: 100
-        },
-        toolbox: {
-            show: true,
-            feature: {
-                mark: {
-                    show: true
-                },
-                dataZoom: {
-                    show: true
-                },
-                dataView: {
-                    show: true,
-                    readOnly: false
-                },
-                restore: {
-                    show: true
-                },
-                saveAsImage: {
-                    show: true
-                }
-            }
-        },
-        xAxis: [{
-            type: 'value',
-            scale: true
-        }],
-        yAxis: [{
-            type: 'value',
-            scale: true
-        }],
-        series: series,
-    };
-    return option;
-}
-
-module.exports = {
-    chart: scatter,
-}
-
-},{"./interpolation":4}],10:[function(require,module,exports){
-var scatter3D = function(ecParams) {
-    var data = [];
-    for (var i = 0; i < ecParams.xAxis.length; ++i) {
-        data.push([ecParams.xAxis[i], ecParams.yAxis[i], ecParams.zAxis[i]]);
-    }
-    var option = {
-        title: {
-            text: ecParams.title,
-        },
-        legend: {
-            data: ecParams.title,
-        },
-        xAxis3D: {
-            type: 'value',
-        },
-        yAxis3D: {
-            type: 'value',
-        },
-        zAxis3D: {
-            type: 'value',
-        },
-
-        grid3D: {
-            viewControl: {
-                // projection: 'orthographic'
-            }
-        },
-        series: [{
-            type: 'scatter3D',
-            data: data,
-        }]
-    };
-    return option;
-}
-
-module.exports = {
-    chart: scatter3D,
-}
-
-},{}],11:[function(require,module,exports){
-var theme = {
-    backgroundColor: 'rgba(36,46,46,0.5)',
-    color: ['#2098d1', 'rgb(102, 166, 30)', '#ff0000', 'rgb(152, 78, 163)', 'rgb(0, 210, 213)',
-        'rgb(255, 127, 0)', 'rgb(157, 141, 0)', 'rgb(127, 128, 205)', 'rgb(179, 233, 0)', 'rgb(255, 0, 41)',
-        'rgb(102, 166, 30)', 'rgb(247, 129, 191)', '#8dd3c7', '#bebada', '#fb8072',
-        '#6699FF', '#ff6666', '#3cb371', '#b8860b', '#fdb462'
-    ],
-    title: {
-        x: 'left',
-        y: 'top',
-        backgroundColor: 'rgba(46,46,46,0.9)',
-        borderColor: '#ccc',
-        borderWidth: 0,
-        padding: 5,
-
-        itemGap: 10,
-        textStyle: {
-            fontSize: 18,
-            fontWeight: 'bolder',
-            color: 'rgba(255, 255, 255)'
-        },
-        subtextStyle: {
-            color: '#aaa'
-        }
-    },
-
-    legend: {
-        orient: 'horizontal',
-        x: 'center',
-        y: 'top',
-        backgroundColor: 'rgba(46,46,46,0.9)',
-        borderColor: '#171717',
-        borderWidth: 0,
-        padding: 9,
-        itemGap: 10,
-        itemWidth: 20,
-        itemHeight: 14,
-        textStyle: {
-            color: 'rgba(255, 255, 255)'
-        }
-    },
-    dataRange: {
-        orient: 'vertical',
-        x: 'left',
-        y: 'bottom',
-        backgroundColor: 'rgba(46,46,46,0.9)',
-        borderColor: '#ccc',
-        borderWidth: 0,
-        padding: 5,
-
-        itemGap: 10,
-        itemWidth: 20,
-        itemHeight: 14,
-        splitNumber: 5,
-        color: ['#1e90ff', '#f0ffff'],
-        textStyle: {
-            color: 'rgba(255, 255, 255)'
-        }
-    },
-
-    toolbox: {
-        orient: 'horizontal',
-        x: 'right',
-        y: 'top',
-        iconStyle: {
-            color: ['#1e90ff', '#22bb22', '#4b0082', '#d2691e'],
-        },
-        backgroundColor: 'rgba(46,46,46,0.9)',
-
-    },
-
-
-    tooltip: {
-        trigger: 'item',
-        showDelay: 20,
-        hideDelay: 100,
-        transitionDuration: 0.4,
-        backgroundColor: 'rgba(0,0,0,0.7)',
-        borderColor: '#333',
-        borderRadius: 4,
-        borderWidth: 0,
-        padding: 5,
-        axisPointer: {
-            type: 'line',
-            lineStyle: {
-                color: '#48b',
-                width: 2,
-                type: 'solid'
-            },
-            shadowStyle: {
-                width: 'auto',
-                color: 'rgba(150,150,150,0.3)'
-            }
-        },
-        textStyle: {
-            color: '#fff'
-        }
-    },
-    dataZoom: {
-        orient: 'horizontal',
-        backgroundColor: 'rgba(0,0,0,0)',
-        dataBackgroundColor: 'rgba(32, 152, 180, 1)',
-        fillerColor: 'rgba(144,197,237,0.2)',
-        handleColor: 'rgba(70,130,180,0.8)'
-    },
-
-    grid: {
-        x: 80,
-        y: 60,
-        x2: 80,
-        y2: 60,
-        borderWidth: 1,
-        borderColor: '#171717'
-    },
-
-    categoryAxis: {
-        position: 'bottom',
-        nameLocation: 'end',
-        boundaryGap: true,
-        axisLine: {
-            show: true,
-            lineStyle: {
-                color: 'rgba(40, 175, 67, 0.66)',
-                width: 1,
-                type: 'solid'
-            }
-        },
-        axisTick: {
-            show: true,
-            interval: 'auto',
-            // onGap: null,
-            inside: false,
-            length: 5,
-            lineStyle: {
-                color: 'rgba(255, 255, 255, 0.5)',
-                width: 1
-            }
-        },
-        axisLabel: {
-            show: true,
-            interval: 'auto',
-            rotate: 0,
-            margin: 8,
-
-            textStyle: {
-                color: 'rgba(255, 255, 255, 0.5)'
-            }
-        },
-        splitLine: {
-            show: true,
-            // onGap: null,
-            lineStyle: {
-                color: ['rgba(255, 255, 255, 0.5)'],
-                width: 0.3,
-                type: 'solid'
-            }
-        },
-        splitArea: {
-            show: false,
-            // onGap: null,
-            areaStyle: {
-                color: ['rgba(250,250,250,0.3)', 'rgba(200,200,200,0.3)']
-            }
-        }
-    },
-
-
-    valueAxis: {
-        position: 'left',
-        nameLocation: 'end',
-        nameTextStyle: {},
-        boundaryGap: [0, 0],
-        splitNumber: 5,
-        axisLine: {
-            show: true,
-            lineStyle: {
-                color: 'rgba(40, 175, 67, 0.66)',
-                width: 0.9,
-                type: 'dashed'
-            }
-        },
-        axisTick: {
-            show: false,
-            inside: false,
-            length: 5,
-            lineStyle: {
-                color: 'rgba(255, 255, 255)',
-                width: 1
-            }
-        },
-        axisLabel: {
-            show: true,
-            rotate: 0,
-            margin: 8,
-            // formatter: null,
-            textStyle: {
-                color: 'rgba(255, 255, 255)'
-            }
-        },
-        splitLine: {
-            show: true,
-            lineStyle: {
-                color: ['rgba(255, 255, 255, 0.5)'],
-                width: 0.3,
-                type: 'solid'
-            }
-        },
-        splitArea: {
-            show: false,
-            areaStyle: {
-                color: ['rgba(250,250,250,0.3)', 'rgba(200,200,200,0.3)']
-            }
-        }
-    },
-
-    polar: {
-        center: ['50%', '50%'],
-        radius: '75%',
-        startAngle: 90,
-        splitNumber: 5,
-        name: {
-            show: true,
-            textStyle: {
-                color: '#333'
-            }
-        },
-        axisLine: {
-            show: true,
-            lineStyle: {
-                color: '#ccc',
-                width: 1,
-                type: 'solid'
-            }
-        },
-        axisLabel: {
-            show: false,
-            textStyle: {
-                color: '#333'
-            }
-        },
-        splitArea: {
-            show: true,
-            areaStyle: {
-                color: ['rgba(250,250,250,0.3)', 'rgba(200,200,200,0.3)']
-            }
-        },
-        splitLine: {
-            show: true,
-            lineStyle: {
-                width: 1,
-                color: '#ccc'
-            }
-        }
-    },
-
-
-    bar: {
-        barMinHeight: 0,
-        // barWidth: null,        
-        barGap: '30%',
-        barCategoryGap: '20%',
-        itemStyle: {
-            normal: {
-                barBorderColor: '#171717',
-                barBorderRadius: 0,
-                barBorderWidth: 0.3,
-                label: {
-                    show: false
-
-                }
-            },
-            emphasis: {
-                barBorderColor: 'rgba(0,0,0,0)',
-                barBorderRadius: 0,
-                barBorderWidth: 1,
-                label: {
-                    show: false
-
-                }
-            }
-        }
-    },
-
-
-    line: {
-        symbol: 'circle',
-        itemStyle: {
-            normal: {
-
-                label: {
-                    show: false
-                },
-                lineStyle: {
-                    width: 2,
-                    type: 'solid',
-                    shadowColor: 'rgba(0,0,0,0)',
-                    shadowBlur: 5,
-                    shadowOffsetX: 3,
-                    shadowOffsetY: 3
-                }
-            },
-            emphasis: {
-
-                label: {
-                    show: false
-
-                }
-            }
-        },
-        //smooth : false,
-        //symbol: null,         
-        symbolSize: 2,
-        //symbolRotate : null,  
-        showAllSymbol: false
-    },
-
-
-    k: {
-        // barWidth : null          
-        // barMaxWidth : null       
-        itemStyle: {
-            normal: {
-                color: '#fff',
-                color0: '#00aa11',
-                lineStyle: {
-                    width: 1,
-                    color: '#ff3200',
-                    color0: '#00aa11'
-                }
-            },
-            emphasis: {}
-        }
-    },
-
-
-    scatter: {
-        //symbol: null,     
-        symbolSize: 4,
-        //symbolRotate : null,  
-        large: false,
-        largeThreshold: 2000,
-        itemStyle: {
-            normal: {
-                label: {
-                    show: false
-
-                }
-            },
-            emphasis: {
-
-                label: {
-                    show: false
-                }
-            }
-        }
-    },
-
-    radar: {
-        itemStyle: {
-            normal: {
-                label: {
-                    show: false
-                },
-                lineStyle: {
-                    width: 2,
-                    type: 'solid'
-                }
-            },
-            emphasis: {
-                label: {
-                    show: false
-                }
-            }
-        },
-        //symbol: null,         
-        symbolSize: 2
-        //symbolRotate : null,  
-    },
-
-
-    pie: {
-        center: ['50%', '50%'],
-        radius: [0, '75%'],
-        clockWise: false,
-        startAngle: 90,
-        minAngle: 0,
-        selectedOffset: 10,
-        itemStyle: {
-            normal: {
-                borderColor: '#fff',
-                borderWidth: 1,
-                label: {
-                    show: true,
-                    position: 'outer'
-                    // textStyle: null    
-                },
-                labelLine: {
-                    show: true,
-                    length: 20,
-                    lineStyle: {
-                        width: 1,
-                        type: 'solid'
-                    }
-                }
-            },
-            emphasis: {
-                borderColor: 'rgba(0,0,0,0)',
-                borderWidth: 1,
-                label: {
-                    show: false
-                    // position: 'outer'
-                    // textStyle: null   
-                },
-                labelLine: {
-                    show: false,
-                    length: 20,
-                    lineStyle: {
-                        width: 1,
-                        type: 'solid'
-                    }
-                }
-            }
-        }
-    },
-    chord: {
-        radius: ['65%', '75%'],
-        center: ['50%', '50%'],
-        padding: 2,
-        sort: 'none', // can be 'none', 'ascending', 'descending'
-        sortSub: 'none', // can be 'none', 'ascending', 'descending'
-        startAngle: 90,
-        clockWise: false,
-        showScale: false,
-        showScaleText: false,
-        itemStyle: {
-            normal: {
-                label: {
-                    show: true
-                    // textStyle: null    
-                },
-                lineStyle: {
-                    width: 0,
-                    color: '#000'
-                },
-                chordStyle: {
-                    lineStyle: {
-                        width: 1,
-                        color: '#666'
-                    }
-                }
-            },
-            emphasis: {
-                lineStyle: {
-                    width: 0,
-                    color: '#000'
-                },
-                chordStyle: {
-                    lineStyle: {
-                        width: 2,
-                        color: '#333'
-                    }
-                }
-            }
-        }
-    },
-
-    island: {
-        r: 15,
-        calculateStep: 0.1
-    },
-
-    textStyle: {
-        decoration: 'none',
-        fontFamily: 'quicksand, sans-serif',
-        fontFamily2: 'rationale, sans-serif',
-        fontSize: 13,
-        fontStyle: 'normal',
-        fontWeight: 'normal'
-    },
-}
-
-module.exports = {
-    theme: theme,
-}
-},{}],12:[function(require,module,exports){
 /**
  * math.js
  * https://github.com/josdejong/mathjs
@@ -1692,73 +53,4 @@ module.exports = {
  * Dual licensed under the MIT or GPL Version 2 licenses.
  **/
 !function(e){"use strict";var n=function(e){return.5*(Math.exp(e)+Math.exp(-e))},i=function(e){return.5*(Math.exp(e)-Math.exp(-e))},u=function(){throw SyntaxError("Invalid Param")};function c(e,t){var r=Math.abs(e),n=Math.abs(t);return 0===e?Math.log(n):0===t?Math.log(r):r<3e3&&n<3e3?.5*Math.log(e*e+t*t):Math.log(e/Math.cos(Math.atan2(t,e)))}var o=function(e,t){var r={re:0,im:0};if(null==e)r.re=r.im=0;else if(void 0!==t)r.re=e,r.im=t;else switch(typeof e){case"object":if("im"in e&&"re"in e)r.re=e.re,r.im=e.im;else if("abs"in e&&"arg"in e){if(!Number.isFinite(e.abs)&&Number.isFinite(e.arg))return f.INFINITY;r.re=e.abs*Math.cos(e.arg),r.im=e.abs*Math.sin(e.arg)}else if("r"in e&&"phi"in e){if(!Number.isFinite(e.r)&&Number.isFinite(e.phi))return f.INFINITY;r.re=e.r*Math.cos(e.phi),r.im=e.r*Math.sin(e.phi)}else 2===e.length?(r.re=e[0],r.im=e[1]):u();break;case"string":r.im=r.re=0;var n=e.match(/\d+\.?\d*e[+-]?\d+|\d+\.?\d*|\.\d+|./g),i=1,o=0;null===n&&u();for(var a=0;a<n.length;a++){var s=n[a];" "===s||"\t"===s||"\n"===s||("+"===s?i++:"-"===s?o++:("i"===s||"I"===s?(i+o===0&&u()," "===n[a+1]||isNaN(n[a+1])?r.im+=parseFloat((o%2?"-":"")+"1"):(r.im+=parseFloat((o%2?"-":"")+n[a+1]),a++)):((i+o===0||isNaN(s))&&u(),"i"===n[a+1]||"I"===n[a+1]?(r.im+=parseFloat((o%2?"-":"")+s),a++):r.re+=parseFloat((o%2?"-":"")+s)),i=o=0))}0<i+o&&u();break;case"number":r.im=0,r.re=e;break;default:u()}return isNaN(r.re)||isNaN(r.im),r};function f(e,t){if(!(this instanceof f))return new f(e,t);var r=o(e,t);this.re=r.re,this.im=r.im}f.prototype={re:0,im:0,sign:function(){var e=this.abs();return new f(this.re/e,this.im/e)},add:function(e,t){var r=new f(e,t);return this.isInfinite()&&r.isInfinite()?f.NAN:this.isInfinite()||r.isInfinite()?f.INFINITY:new f(this.re+r.re,this.im+r.im)},sub:function(e,t){var r=new f(e,t);return this.isInfinite()&&r.isInfinite()?f.NAN:this.isInfinite()||r.isInfinite()?f.INFINITY:new f(this.re-r.re,this.im-r.im)},mul:function(e,t){var r=new f(e,t);return this.isInfinite()&&r.isZero()||this.isZero()&&r.isInfinite()?f.NAN:this.isInfinite()||r.isInfinite()?f.INFINITY:0===r.im&&0===this.im?new f(this.re*r.re,0):new f(this.re*r.re-this.im*r.im,this.re*r.im+this.im*r.re)},div:function(e,t){var r=new f(e,t);if(this.isZero()&&r.isZero()||this.isInfinite()&&r.isInfinite())return f.NAN;if(this.isInfinite()||r.isZero())return f.INFINITY;if(this.isZero()||r.isInfinite())return f.ZERO;e=this.re,t=this.im;var n,i,o=r.re,a=r.im;return 0===a?new f(e/o,t/o):Math.abs(o)<Math.abs(a)?new f((e*(i=o/a)+t)/(n=o*i+a),(t*i-e)/n):new f((e+t*(i=a/o))/(n=a*i+o),(t-e*i)/n)},pow:function(e,t){var r=new f(e,t);if(e=this.re,t=this.im,r.isZero())return f.ONE;if(0===r.im){if(0===t&&0<=e)return new f(Math.pow(e,r.re),0);if(0===e)switch((r.re%4+4)%4){case 0:return new f(Math.pow(t,r.re),0);case 1:return new f(0,Math.pow(t,r.re));case 2:return new f(-Math.pow(t,r.re),0);case 3:return new f(0,-Math.pow(t,r.re))}}if(0===e&&0===t&&0<r.re&&0<=r.im)return f.ZERO;var n=Math.atan2(t,e),i=c(e,t);return e=Math.exp(r.re*i-r.im*n),t=r.im*i+r.re*n,new f(e*Math.cos(t),e*Math.sin(t))},sqrt:function(){var e,t,r=this.re,n=this.im,i=this.abs();if(0<=r){if(0===n)return new f(Math.sqrt(r),0);e=.5*Math.sqrt(2*(i+r))}else e=Math.abs(n)/Math.sqrt(2*(i-r));return t=r<=0?.5*Math.sqrt(2*(i-r)):Math.abs(n)/Math.sqrt(2*(i+r)),new f(e,n<0?-t:t)},exp:function(){var e=Math.exp(this.re);return this.im,new f(e*Math.cos(this.im),e*Math.sin(this.im))},expm1:function(){var e=this.re,t=this.im;return new f(Math.expm1(e)*Math.cos(t)+function(e){var t=Math.PI/4;if(e<-t||t<e)return Math.cos(e)-1;var r=e*e;return r*(r*(1/24+r*(-1/720+r*(1/40320+r*(-1/3628800+r*(1/4790014600+r*(-1/87178291200+r*(1/20922789888e3)))))))-.5)}(t),Math.exp(e)*Math.sin(t))},log:function(){var e=this.re,t=this.im;return new f(c(e,t),Math.atan2(t,e))},abs:function(){return e=this.re,t=this.im,r=Math.abs(e),n=Math.abs(t),r<3e3&&n<3e3?Math.sqrt(r*r+n*n):(r<n?(r=n,n=e/t):n=t/e,r*Math.sqrt(1+n*n));var e,t,r,n},arg:function(){return Math.atan2(this.im,this.re)},sin:function(){var e=this.re,t=this.im;return new f(Math.sin(e)*n(t),Math.cos(e)*i(t))},cos:function(){var e=this.re,t=this.im;return new f(Math.cos(e)*n(t),-Math.sin(e)*i(t))},tan:function(){var e=2*this.re,t=2*this.im,r=Math.cos(e)+n(t);return new f(Math.sin(e)/r,i(t)/r)},cot:function(){var e=2*this.re,t=2*this.im,r=Math.cos(e)-n(t);return new f(-Math.sin(e)/r,i(t)/r)},sec:function(){var e=this.re,t=this.im,r=.5*n(2*t)+.5*Math.cos(2*e);return new f(Math.cos(e)*n(t)/r,Math.sin(e)*i(t)/r)},csc:function(){var e=this.re,t=this.im,r=.5*n(2*t)-.5*Math.cos(2*e);return new f(Math.sin(e)*n(t)/r,-Math.cos(e)*i(t)/r)},asin:function(){var e=this.re,t=this.im,r=new f(t*t-e*e+1,-2*e*t).sqrt(),n=new f(r.re-t,r.im+e).log();return new f(n.im,-n.re)},acos:function(){var e=this.re,t=this.im,r=new f(t*t-e*e+1,-2*e*t).sqrt(),n=new f(r.re-t,r.im+e).log();return new f(Math.PI/2-n.im,n.re)},atan:function(){var e=this.re,t=this.im;if(0===e){if(1===t)return new f(0,1/0);if(-1===t)return new f(0,-1/0)}var r=e*e+(1-t)*(1-t),n=new f((1-t*t-e*e)/r,-2*e/r).log();return new f(-.5*n.im,.5*n.re)},acot:function(){var e=this.re,t=this.im;if(0===t)return new f(Math.atan2(1,e),0);var r=e*e+t*t;return 0!==r?new f(e/r,-t/r).atan():new f(0!==e?e/0:0,0!==t?-t/0:0).atan()},asec:function(){var e=this.re,t=this.im;if(0===e&&0===t)return new f(0,1/0);var r=e*e+t*t;return 0!==r?new f(e/r,-t/r).acos():new f(0!==e?e/0:0,0!==t?-t/0:0).acos()},acsc:function(){var e=this.re,t=this.im;if(0===e&&0===t)return new f(Math.PI/2,1/0);var r=e*e+t*t;return 0!==r?new f(e/r,-t/r).asin():new f(0!==e?e/0:0,0!==t?-t/0:0).asin()},sinh:function(){var e=this.re,t=this.im;return new f(i(e)*Math.cos(t),n(e)*Math.sin(t))},cosh:function(){var e=this.re,t=this.im;return new f(n(e)*Math.cos(t),i(e)*Math.sin(t))},tanh:function(){var e=2*this.re,t=2*this.im,r=n(e)+Math.cos(t);return new f(i(e)/r,Math.sin(t)/r)},coth:function(){var e=2*this.re,t=2*this.im,r=n(e)-Math.cos(t);return new f(i(e)/r,-Math.sin(t)/r)},csch:function(){var e=this.re,t=this.im,r=Math.cos(2*t)-n(2*e);return new f(-2*i(e)*Math.cos(t)/r,2*n(e)*Math.sin(t)/r)},sech:function(){var e=this.re,t=this.im,r=Math.cos(2*t)+n(2*e);return new f(2*n(e)*Math.cos(t)/r,-2*i(e)*Math.sin(t)/r)},asinh:function(){var e=this.im;this.im=-this.re,this.re=e;var t=this.asin();return this.re=-this.im,this.im=e,e=t.re,t.re=-t.im,t.im=e,t},acosh:function(){var e=this.acos();if(e.im<=0){var t=e.re;e.re=-e.im,e.im=t}else{t=e.im;e.im=-e.re,e.re=t}return e},atanh:function(){var e=this.re,t=this.im,r=1<e&&0===t,n=1-e,i=1+e,o=n*n+t*t,a=0!==o?new f((i*n-t*t)/o,(t*n+i*t)/o):new f(-1!==e?e/0:0,0!==t?t/0:0),s=a.re;return a.re=c(a.re,a.im)/2,a.im=Math.atan2(a.im,s)/2,r&&(a.im=-a.im),a},acoth:function(){var e=this.re,t=this.im;if(0===e&&0===t)return new f(0,Math.PI/2);var r=e*e+t*t;return 0!==r?new f(e/r,-t/r).atanh():new f(0!==e?e/0:0,0!==t?-t/0:0).atanh()},acsch:function(){var e=this.re,t=this.im;if(0===t)return new f(0!==e?Math.log(e+Math.sqrt(e*e+1)):1/0,0);var r=e*e+t*t;return 0!==r?new f(e/r,-t/r).asinh():new f(0!==e?e/0:0,0!==t?-t/0:0).asinh()},asech:function(){var e=this.re,t=this.im;if(this.isZero())return f.INFINITY;var r=e*e+t*t;return 0!==r?new f(e/r,-t/r).acosh():new f(0!==e?e/0:0,0!==t?-t/0:0).acosh()},inverse:function(){if(this.isZero())return f.INFINITY;if(this.isInfinite())return f.ZERO;var e=this.re,t=this.im,r=e*e+t*t;return new f(e/r,-t/r)},conjugate:function(){return new f(this.re,-this.im)},neg:function(){return new f(-this.re,-this.im)},ceil:function(e){return e=Math.pow(10,e||0),new f(Math.ceil(this.re*e)/e,Math.ceil(this.im*e)/e)},floor:function(e){return e=Math.pow(10,e||0),new f(Math.floor(this.re*e)/e,Math.floor(this.im*e)/e)},round:function(e){return e=Math.pow(10,e||0),new f(Math.round(this.re*e)/e,Math.round(this.im*e)/e)},equals:function(e,t){var r=new f(e,t);return Math.abs(r.re-this.re)<=f.EPSILON&&Math.abs(r.im-this.im)<=f.EPSILON},clone:function(){return new f(this.re,this.im)},toString:function(){var e=this.re,t=this.im,r="";return this.isNaN()?"NaN":this.isZero()?"0":this.isInfinite()?"Infinity":(0!==e&&(r+=e),0!==t&&(0!==e?r+=t<0?" - ":" + ":t<0&&(r+="-"),1!==(t=Math.abs(t))&&(r+=t),r+="i"),r||"0")},toVector:function(){return[this.re,this.im]},valueOf:function(){return 0===this.im?this.re:null},isNaN:function(){return isNaN(this.re)||isNaN(this.im)},isZero:function(){return!(0!==this.re&&-0!==this.re||0!==this.im&&-0!==this.im)},isFinite:function(){return isFinite(this.re)&&isFinite(this.im)},isInfinite:function(){return!(this.isNaN()||this.isFinite())}},f.ZERO=new f(0,0),f.ONE=new f(1,0),f.I=new f(0,1),f.PI=new f(Math.PI,0),f.E=new f(Math.E,0),f.INFINITY=new f(1/0,1/0),f.NAN=new f(NaN,NaN),f.EPSILON=1e-16,void 0===(a=function(){return f}.apply(r,[]))||(t.exports=a)}()},function(e,t,r){"use strict";e.exports=[r(91),r(550)]},function(e,t,r){"use strict";t.name="chain",t.factory=function(t,e,r,n){return n("chain",{"":function(){return new t.Chain},any:function(e){return new t.Chain(e)}})}},function(e,u,t){"use strict";var c=t(5);u.format=function(e,t){if("function"==typeof t)return t(e);if(!e.isFinite())return e.isNaN()?"NaN":e.gt(0)?"Infinity":"-Infinity";var r="auto",n=void 0;switch(void 0!==t&&(t.notation&&(r=t.notation),"number"==typeof t?n=t:t.precision&&(n=t.precision)),r){case"fixed":return u.toFixed(e,n);case"exponential":return u.toExponential(e,n);case"auto":if(t&&t.exponential&&(void 0!==t.exponential.lower||void 0!==t.exponential.upper)){var i=c.map(t,function(e){return e});return(i.exponential=void 0)!==t.exponential.lower&&(i.lowerExp=Math.round(Math.log(t.exponential.lower)/Math.LN10)),void 0!==t.exponential.upper&&(i.upperExp=Math.round(Math.log(t.exponential.upper)/Math.LN10)),console.warn("Deprecation warning: Formatting options exponential.lower and exponential.upper (minimum and maximum value) are replaced with exponential.lowerExp and exponential.upperExp (minimum and maximum exponent) since version 4.0.0. Replace "+JSON.stringify(t)+" with "+JSON.stringify(i)),u.format(e,i)}var o=t&&void 0!==t.lowerExp?t.lowerExp:-3,a=t&&void 0!==t.upperExp?t.upperExp:5;if(e.isZero())return"0";var s=e.logarithm();return(s.gte(o)&&s.lt(a)?e.toSignificantDigits(n).toFixed():u.toExponential(e,n)).replace(/((\.\d*?)(0+))($|e)/,function(){var e=arguments[2],t=arguments[4];return"."!==e?e+t:t});default:throw new Error('Unknown notation "'+r+'". Choose "auto", "exponential", or "fixed".')}},u.toExponential=function(e,t){return void 0!==t?e.toExponential(t-1):e.toExponential()},u.toFixed=function(e,t){return e.toFixed(t)}},function(e,t,r){"use strict";var u=r(9).format,c=r(5).lazy;t.name="Chain",t.path="type",t.factory=function(t,e,r,n,i){function o(e){if(!(this instanceof o))throw new SyntaxError("Constructor must be called with the new operator");t.isChain(e)?this.value=e.value:this.value=e}function a(e,t){"function"==typeof t&&(o.prototype[e]=s(t))}function s(r){return function(){for(var e=[this.value],t=0;t<arguments.length;t++)e[t+1]=arguments[t];return new o(r.apply(r,e))}}return o.prototype.type="Chain",o.prototype.isChain=!0,o.prototype.done=function(){return this.value},o.prototype.valueOf=function(){return this.value},o.prototype.toString=function(){return u(this.value)},o.prototype.toJSON=function(){return{mathjs:"Chain",value:this.value}},o.fromJSON=function(e){return new o(e.value)},o.createProxy=function(e,t){if("string"==typeof e)a(e,t);else for(var r in e)e.hasOwnProperty(r)&&a(r,e[r])},o.createProxy(i),i.on("import",function(e,t,r){var n;void 0===r&&(n=t,c(o.prototype,e,function(){var e=n();if("function"==typeof e)return s(e)}))}),o},t.math=!0,t.lazy=!1},function(e,t,r){"use strict";e.exports=[r(555),r(553)]},function(e,t,r){"use strict";var o=r(0);t.name="boolean",t.factory=function(e,t,r,n){var i=n("bool",{"":function(){return!1},boolean:function(e){return e},number:function(e){return!!e},null:function(e){return!1},BigNumber:function(e){return!e.isZero()},string:function(e){var t=e.toLowerCase();if("true"===t)return!0;if("false"===t)return!1;var r=Number(e);if(""!==e&&!isNaN(r))return!!r;throw new Error('Cannot convert "'+e+'" to a boolean')},"Array | Matrix":function(e){return o(e,i)}});return i}},function(e,t,r){"use strict";var o=r(0);t.name="bignumber",t.factory=function(t,e,r,n){var i=n("bignumber",{"":function(){return new t.BigNumber(0)},number:function(e){return new t.BigNumber(e+"")},string:function(e){return new t.BigNumber(e)},BigNumber:function(e){return e},Fraction:function(e){return new t.BigNumber(e.n).div(e.d)},null:function(e){return new t.BigNumber(0)},"Array | Matrix":function(e){return o(e,i)}});return i.toTex={0:"0",1:"\\left(${args[0]}\\right)"},i}},function(Ie,ke,Pe){var Re;!function(e){"use strict";var l,I,a,s=9e15,d=1e9,v="0123456789abcdef",n="2.3025850929940456840179914546843642076011014886287729760333279009675726096773524802359972050895982983419677840422862486334095254650828067566662873690987816894829072083255546808437998948262331985283935053089653777326288461633662222876982198867465436674744042432743651550489343149393914796194044002221051017141748003688084012647080685567743216228355220114804663715659121373450747856947683463616792101806445070648000277502684916746550586856935673420670581136429224554405758925724208241314695689016758940256776311356919292033376587141660230105703089634572075440370847469940168269282808481184289314848524948644871927809676271275775397027668605952496716674183485704422507197965004714951050492214776567636938662976979522110718264549734772662425709429322582798502585509785265383207606726317164309505995087807523710333101197857547331541421808427543863591778117054309827482385045648019095610299291824318237525357709750539565187697510374970888692180205189339507238539205144634197265287286965110862571492198849978748873771345686209167058",i="3.1415926535897932384626433832795028841971693993751058209749445923078164062862089986280348253421170679821480865132823066470938446095505822317253594081284811174502841027019385211055596446229489549303819644288109756659334461284756482337867831652712019091456485669234603486104543266482133936072602491412737245870066063155881748815209209628292540917153643678925903600113305305488204665213841469519415116094330572703657595919530921861173819326117931051185480744623799627495673518857527248912279381830119491298336733624406566430860213949463952247371907021798609437027705392171762931767523846748184676694051320005681271452635608277857713427577896091736371787214684409012249534301465495853710507922796892589235420199561121290219608640344181598136297747713099605187072113499999983729780499510597317328160963185950244594553469083026425223082533446850352619311881710100031378387528865875332083814206171776691473035982534904287554687311595628638823537875937519577818577805321712268066130019278766111959092164201989380952572010654858632789",u={precision:20,rounding:4,modulo:1,toExpNeg:-7,toExpPos:21,minE:-s,maxE:s,crypto:!1},x=!0,c="[DecimalError] ",y=c+"Invalid argument: ",o=c+"Precision limit exceeded",f=c+"crypto unavailable",k=Math.floor,g=Math.pow,p=/^0b([01]+(\.[01]*)?|\.[01]+)(p[+-]?\d+)?$/i,m=/^0x([0-9a-f]+(\.[0-9a-f]*)?|\.[0-9a-f]+)(p[+-]?\d+)?$/i,h=/^0o([0-7]+(\.[0-7]*)?|\.[0-7]+)(p[+-]?\d+)?$/i,b=/^(\d+(\.\d*)?|\.\d+)(e[+-]?\d+)?$/i,P=1e7,R=7,w=n.length-1,N=i.length-1,M={name:"[object Decimal]"};function E(e){var t,r,n,i=e.length-1,o="",a=e[0];if(0<i){for(o+=a,t=1;t<i;t++)n=e[t]+"",(r=R-n.length)&&(o+=D(r)),o+=n;a=e[t],(r=R-(n=a+"").length)&&(o+=D(r))}else if(0===a)return"0";for(;a%10==0;)a/=10;return o+a}function A(e,t,r){if(e!==~~e||e<t||r<e)throw Error(y+e)}function S(e,t,r,n){var i,o,a,s;for(o=e[0];10<=o;o/=10)--t;return--t<0?(t+=R,i=0):(i=Math.ceil((t+1)/R),t%=R),o=g(10,R-t),s=e[i]%o|0,null==n?t<3?(0==t?s=s/100|0:1==t&&(s=s/10|0),a=r<4&&99999==s||3<r&&49999==s||5e4==s||0==s):a=(r<4&&s+1==o||3<r&&s+1==o/2)&&(e[i+1]/o/100|0)==g(10,t-2)-1||(s==o/2||0==s)&&0==(e[i+1]/o/100|0):t<4?(0==t?s=s/1e3|0:1==t?s=s/100|0:2==t&&(s=s/10|0),a=(n||r<4)&&9999==s||!n&&3<r&&4999==s):a=((n||r<4)&&s+1==o||!n&&3<r&&s+1==o/2)&&(e[i+1]/o/1e3|0)==g(10,t-3)-1,a}function O(e,t,r){for(var n,i,o=[0],a=0,s=e.length;a<s;){for(i=o.length;i--;)o[i]*=t;for(o[0]+=v.indexOf(e.charAt(a++)),n=0;n<o.length;n++)o[n]>r-1&&(void 0===o[n+1]&&(o[n+1]=0),o[n+1]+=o[n]/r|0,o[n]%=r)}return o.reverse()}M.absoluteValue=M.abs=function(){var e=new this.constructor(this);return e.s<0&&(e.s=1),U(e)},M.ceil=function(){return U(new this.constructor(this),this.e+1,2)},M.comparedTo=M.cmp=function(e){var t,r,n,i,o=this.d,a=(e=new this.constructor(e)).d,s=this.s,u=e.s;if(!o||!a)return s&&u?s!==u?s:o===a?0:!o^s<0?1:-1:NaN;if(!o[0]||!a[0])return o[0]?s:a[0]?-u:0;if(s!==u)return s;if(this.e!==e.e)return this.e>e.e^s<0?1:-1;for(t=0,r=(n=o.length)<(i=a.length)?n:i;t<r;++t)if(o[t]!==a[t])return o[t]>a[t]^s<0?1:-1;return n===i?0:i<n^s<0?1:-1},M.cosine=M.cos=function(){var e,t,r=this,n=r.constructor;return r.d?r.d[0]?(e=n.precision,t=n.rounding,n.precision=e+Math.max(r.e,r.sd())+R,n.rounding=1,r=function(e,t){var r,n,i=t.d.length;i<32?(r=Math.ceil(i/3),n=Math.pow(4,-r).toString()):(r=16,n="2.3283064365386962890625e-10");e.precision+=r,t=Z(e,1,t.times(n),new e(1));for(var o=r;o--;){var a=t.times(t);t=a.times(a).minus(a).times(8).plus(1)}return e.precision-=r,t}(n,V(n,r)),n.precision=e,n.rounding=t,U(2==a||3==a?r.neg():r,e,t,!0)):new n(1):new n(NaN)},M.cubeRoot=M.cbrt=function(){var e,t,r,n,i,o,a,s,u,c,f=this,l=f.constructor;if(!f.isFinite()||f.isZero())return new l(f);for(x=!1,(o=f.s*Math.pow(f.s*f,1/3))&&Math.abs(o)!=1/0?n=new l(o.toString()):(r=E(f.d),(o=((e=f.e)-r.length+1)%3)&&(r+=1==o||-2==o?"0":"00"),o=Math.pow(r,1/3),e=k((e+1)/3)-(e%3==(e<0?-1:2)),(n=new l(r=o==1/0?"5e"+e:(r=o.toExponential()).slice(0,r.indexOf("e")+1)+e)).s=f.s),a=(e=l.precision)+3;;)if(c=(u=(s=n).times(s).times(s)).plus(f),n=T(c.plus(f).times(s),c.plus(u),a+2,1),E(s.d).slice(0,a)===(r=E(n.d)).slice(0,a)){if("9999"!=(r=r.slice(a-3,a+1))&&(i||"4999"!=r)){+r&&(+r.slice(1)||"5"!=r.charAt(0))||(U(n,e+1,1),t=!n.times(n).times(n).eq(f));break}if(!i&&(U(s,e+1,0),s.times(s).times(s).eq(f))){n=s;break}a+=4,i=1}return x=!0,U(n,e,l.rounding,t)},M.decimalPlaces=M.dp=function(){var e,t=this.d,r=NaN;if(t){if(r=((e=t.length-1)-k(this.e/R))*R,e=t[e])for(;e%10==0;e/=10)r--;r<0&&(r=0)}return r},M.dividedBy=M.div=function(e){return T(this,new this.constructor(e))},M.dividedToIntegerBy=M.divToInt=function(e){var t=this.constructor;return U(T(this,new t(e),0,1,1),t.precision,t.rounding)},M.equals=M.eq=function(e){return 0===this.cmp(e)},M.floor=function(){return U(new this.constructor(this),this.e+1,3)},M.greaterThan=M.gt=function(e){return 0<this.cmp(e)},M.greaterThanOrEqualTo=M.gte=function(e){var t=this.cmp(e);return 1==t||0===t},M.hyperbolicCosine=M.cosh=function(){var e,t,r,n,i,o=this,a=o.constructor,s=new a(1);if(!o.isFinite())return new a(o.s?1/0:NaN);if(o.isZero())return s;r=a.precision,n=a.rounding,a.precision=r+Math.max(o.e,o.sd())+4,a.rounding=1,(i=o.d.length)<32?(e=Math.ceil(i/3),t=Math.pow(4,-e).toString()):(e=16,t="2.3283064365386962890625e-10"),o=Z(a,1,o.times(t),new a(1),!0);for(var u,c=e,f=new a(8);c--;)u=o.times(o),o=s.minus(u.times(f.minus(u.times(f))));return U(o,a.precision=r,a.rounding=n,!0)},M.hyperbolicSine=M.sinh=function(){var e,t,r,n,i=this,o=i.constructor;if(!i.isFinite()||i.isZero())return new o(i);if(t=o.precision,r=o.rounding,o.precision=t+Math.max(i.e,i.sd())+4,o.rounding=1,(n=i.d.length)<3)i=Z(o,2,i,i,!0);else{e=16<(e=1.4*Math.sqrt(n))?16:0|e,i=Z(o,2,i=i.times(Math.pow(5,-e)),i,!0);for(var a,s=new o(5),u=new o(16),c=new o(20);e--;)a=i.times(i),i=i.times(s.plus(a.times(u.times(a).plus(c))))}return U(i,o.precision=t,o.rounding=r,!0)},M.hyperbolicTangent=M.tanh=function(){var e,t,r=this.constructor;return this.isFinite()?this.isZero()?new r(this):(e=r.precision,t=r.rounding,r.precision=e+7,r.rounding=1,T(this.sinh(),this.cosh(),r.precision=e,r.rounding=t)):new r(this.s)},M.inverseCosine=M.acos=function(){var e,t=this,r=t.constructor,n=t.abs().cmp(1),i=r.precision,o=r.rounding;return-1!==n?0===n?t.isNeg()?B(r,i,o):new r(0):new r(NaN):t.isZero()?B(r,i+4,o).times(.5):(r.precision=i+6,r.rounding=1,t=t.asin(),e=B(r,i+4,o).times(.5),r.precision=i,r.rounding=o,e.minus(t))},M.inverseHyperbolicCosine=M.acosh=function(){var e,t,r=this,n=r.constructor;return r.lte(1)?new n(r.eq(1)?0:NaN):r.isFinite()?(e=n.precision,t=n.rounding,n.precision=e+Math.max(Math.abs(r.e),r.sd())+4,n.rounding=1,x=!1,r=r.times(r).minus(1).sqrt().plus(r),x=!0,n.precision=e,n.rounding=t,r.ln()):new n(r)},M.inverseHyperbolicSine=M.asinh=function(){var e,t,r=this,n=r.constructor;return!r.isFinite()||r.isZero()?new n(r):(e=n.precision,t=n.rounding,n.precision=e+2*Math.max(Math.abs(r.e),r.sd())+6,n.rounding=1,x=!1,r=r.times(r).plus(1).sqrt().plus(r),x=!0,n.precision=e,n.rounding=t,r.ln())},M.inverseHyperbolicTangent=M.atanh=function(){var e,t,r,n,i=this,o=i.constructor;return i.isFinite()?0<=i.e?new o(i.abs().eq(1)?i.s/0:i.isZero()?i:NaN):(e=o.precision,t=o.rounding,n=i.sd(),Math.max(n,e)<2*-i.e-1?U(new o(i),e,t,!0):(o.precision=r=n-i.e,i=T(i.plus(1),new o(1).minus(i),r+e,1),o.precision=e+4,o.rounding=1,i=i.ln(),o.precision=e,o.rounding=t,i.times(.5))):new o(NaN)},M.inverseSine=M.asin=function(){var e,t,r,n,i=this,o=i.constructor;return i.isZero()?new o(i):(t=i.abs().cmp(1),r=o.precision,n=o.rounding,-1!==t?0===t?((e=B(o,r+4,n).times(.5)).s=i.s,e):new o(NaN):(o.precision=r+6,o.rounding=1,i=i.div(new o(1).minus(i.times(i)).sqrt().plus(1)).atan(),o.precision=r,o.rounding=n,i.times(2)))},M.inverseTangent=M.atan=function(){var e,t,r,n,i,o,a,s,u,c=this,f=c.constructor,l=f.precision,p=f.rounding;if(c.isFinite()){if(c.isZero())return new f(c);if(c.abs().eq(1)&&l+4<=N)return(a=B(f,l+4,p).times(.25)).s=c.s,a}else{if(!c.s)return new f(NaN);if(l+4<=N)return(a=B(f,l+4,p).times(.5)).s=c.s,a}for(f.precision=s=l+10,f.rounding=1,e=r=Math.min(28,s/R+2|0);e;--e)c=c.div(c.times(c).plus(1).sqrt().plus(1));for(x=!1,t=Math.ceil(s/R),n=1,u=c.times(c),a=new f(c),i=c;-1!==e;)if(i=i.times(u),o=a.minus(i.div(n+=2)),i=i.times(u),void 0!==(a=o.plus(i.div(n+=2))).d[t])for(e=t;a.d[e]===o.d[e]&&e--;);return r&&(a=a.times(2<<r-1)),x=!0,U(a,f.precision=l,f.rounding=p,!0)},M.isFinite=function(){return!!this.d},M.isInteger=M.isInt=function(){return!!this.d&&k(this.e/R)>this.d.length-2},M.isNaN=function(){return!this.s},M.isNegative=M.isNeg=function(){return this.s<0},M.isPositive=M.isPos=function(){return 0<this.s},M.isZero=function(){return!!this.d&&0===this.d[0]},M.lessThan=M.lt=function(e){return this.cmp(e)<0},M.lessThanOrEqualTo=M.lte=function(e){return this.cmp(e)<1},M.logarithm=M.log=function(e){var t,r,n,i,o,a,s,u,c=this.constructor,f=c.precision,l=c.rounding;if(null==e)e=new c(10),t=!0;else{if(r=(e=new c(e)).d,e.s<0||!r||!r[0]||e.eq(1))return new c(NaN);t=e.eq(10)}if(r=this.d,this.s<0||!r||!r[0]||this.eq(1))return new c(r&&!r[0]?-1/0:1!=this.s?NaN:r?0:1/0);if(t)if(1<r.length)o=!0;else{for(i=r[0];i%10==0;)i/=10;o=1!==i}if(x=!1,a=$(this,s=f+5),n=t?z(c,s+10):$(e,s),S((u=T(a,n,s,1)).d,i=f,l))do{if(a=$(this,s+=10),n=t?z(c,s+10):$(e,s),u=T(a,n,s,1),!o){+E(u.d).slice(i+1,i+15)+1==1e14&&(u=U(u,f+1,0));break}}while(S(u.d,i+=10,l));return x=!0,U(u,f,l)},M.minus=M.sub=function(e){var t,r,n,i,o,a,s,u,c,f,l,p,m=this,h=m.constructor;if(e=new h(e),!m.d||!e.d)return m.s&&e.s?m.d?e.s=-e.s:e=new h(e.d||m.s!==e.s?m:NaN):e=new h(NaN),e;if(m.s!=e.s)return e.s=-e.s,m.plus(e);if(c=m.d,p=e.d,s=h.precision,u=h.rounding,!c[0]||!p[0]){if(p[0])e.s=-e.s;else{if(!c[0])return new h(3===u?-0:0);e=new h(m)}return x?U(e,s,u):e}if(r=k(e.e/R),f=k(m.e/R),c=c.slice(),o=f-r){for((l=o<0)?(t=c,o=-o,a=p.length):(t=p,r=f,a=c.length),(n=Math.max(Math.ceil(s/R),a)+2)<o&&(o=n,t.length=1),t.reverse(),n=o;n--;)t.push(0);t.reverse()}else{for((l=(n=c.length)<(a=p.length))&&(a=n),n=0;n<a;n++)if(c[n]!=p[n]){l=c[n]<p[n];break}o=0}for(l&&(t=c,c=p,p=t,e.s=-e.s),a=c.length,n=p.length-a;0<n;--n)c[a++]=0;for(n=p.length;o<n;){if(c[--n]<p[n]){for(i=n;i&&0===c[--i];)c[i]=P-1;--c[i],c[n]+=P}c[n]-=p[n]}for(;0===c[--a];)c.pop();for(;0===c[0];c.shift())--r;return c[0]?(e.d=c,e.e=C(c,r),x?U(e,s,u):e):new h(3===u?-0:0)},M.modulo=M.mod=function(e){var t,r=this.constructor;return e=new r(e),!this.d||!e.s||e.d&&!e.d[0]?new r(NaN):!e.d||this.d&&!this.d[0]?U(new r(this),r.precision,r.rounding):(x=!1,9==r.modulo?(t=T(this,e.abs(),0,3,1)).s*=e.s:t=T(this,e,0,r.modulo,1),t=t.times(e),x=!0,this.minus(t))},M.naturalExponential=M.exp=function(){return j(this)},M.naturalLogarithm=M.ln=function(){return $(this)},M.negated=M.neg=function(){var e=new this.constructor(this);return e.s=-e.s,U(e)},M.plus=M.add=function(e){var t,r,n,i,o,a,s,u,c,f,l=this,p=l.constructor;if(e=new p(e),!l.d||!e.d)return l.s&&e.s?l.d||(e=new p(e.d||l.s===e.s?l:NaN)):e=new p(NaN),e;if(l.s!=e.s)return e.s=-e.s,l.minus(e);if(c=l.d,f=e.d,s=p.precision,u=p.rounding,!c[0]||!f[0])return f[0]||(e=new p(l)),x?U(e,s,u):e;if(o=k(l.e/R),n=k(e.e/R),c=c.slice(),i=o-n){for(i<0?(r=c,i=-i,a=f.length):(r=f,n=o,a=c.length),(a=a<(o=Math.ceil(s/R))?o+1:a+1)<i&&(i=a,r.length=1),r.reverse();i--;)r.push(0);r.reverse()}for((a=c.length)-(i=f.length)<0&&(i=a,r=f,f=c,c=r),t=0;i;)t=(c[--i]=c[i]+f[i]+t)/P|0,c[i]%=P;for(t&&(c.unshift(t),++n),a=c.length;0==c[--a];)c.pop();return e.d=c,e.e=C(c,n),x?U(e,s,u):e},M.precision=M.sd=function(e){var t;if(void 0!==e&&e!==!!e&&1!==e&&0!==e)throw Error(y+e);return this.d?(t=q(this.d),e&&this.e+1>t&&(t=this.e+1)):t=NaN,t},M.round=function(){var e=this.constructor;return U(new e(this),this.e+1,e.rounding)},M.sine=M.sin=function(){var e,t,r=this,n=r.constructor;return r.isFinite()?r.isZero()?new n(r):(e=n.precision,t=n.rounding,n.precision=e+Math.max(r.e,r.sd())+R,n.rounding=1,r=function(e,t){var r,n=t.d.length;if(n<3)return Z(e,2,t,t);r=16<(r=1.4*Math.sqrt(n))?16:0|r,t=t.times(Math.pow(5,-r)),t=Z(e,2,t,t);for(var i,o=new e(5),a=new e(16),s=new e(20);r--;)i=t.times(t),t=t.times(o.plus(i.times(a.times(i).minus(s))));return t}(n,V(n,r)),n.precision=e,n.rounding=t,U(2<a?r.neg():r,e,t,!0)):new n(NaN)},M.squareRoot=M.sqrt=function(){var e,t,r,n,i,o,a=this.d,s=this.e,u=this.s,c=this.constructor;if(1!==u||!a||!a[0])return new c(!u||u<0&&(!a||a[0])?NaN:a?this:1/0);for(x=!1,0==(u=Math.sqrt(+this))||u==1/0?(((t=E(a)).length+s)%2==0&&(t+="0"),u=Math.sqrt(t),s=k((s+1)/2)-(s<0||s%2),n=new c(t=u==1/0?"1e"+s:(t=u.toExponential()).slice(0,t.indexOf("e")+1)+s)):n=new c(u.toString()),r=(s=c.precision)+3;;)if(n=(o=n).plus(T(this,o,r+2,1)).times(.5),E(o.d).slice(0,r)===(t=E(n.d)).slice(0,r)){if("9999"!=(t=t.slice(r-3,r+1))&&(i||"4999"!=t)){+t&&(+t.slice(1)||"5"!=t.charAt(0))||(U(n,s+1,1),e=!n.times(n).eq(this));break}if(!i&&(U(o,s+1,0),o.times(o).eq(this))){n=o;break}r+=4,i=1}return x=!0,U(n,s,c.rounding,e)},M.tangent=M.tan=function(){var e,t,r=this,n=r.constructor;return r.isFinite()?r.isZero()?new n(r):(e=n.precision,t=n.rounding,n.precision=e+10,n.rounding=1,(r=r.sin()).s=1,r=T(r,new n(1).minus(r.times(r)).sqrt(),e+10,0),n.precision=e,n.rounding=t,U(2==a||4==a?r.neg():r,e,t,!0)):new n(NaN)},M.times=M.mul=function(e){var t,r,n,i,o,a,s,u,c,f=this.constructor,l=this.d,p=(e=new f(e)).d;if(e.s*=this.s,!(l&&l[0]&&p&&p[0]))return new f(!e.s||l&&!l[0]&&!p||p&&!p[0]&&!l?NaN:l&&p?0*e.s:e.s/0);for(r=k(this.e/R)+k(e.e/R),(u=l.length)<(c=p.length)&&(o=l,l=p,p=o,a=u,u=c,c=a),o=[],n=a=u+c;n--;)o.push(0);for(n=c;0<=--n;){for(t=0,i=u+n;n<i;)s=o[i]+p[n]*l[i-n-1]+t,o[i--]=s%P|0,t=s/P|0;o[i]=(o[i]+t)%P|0}for(;!o[--a];)o.pop();return t?++r:o.shift(),e.d=o,e.e=C(o,r),x?U(e,f.precision,f.rounding):e},M.toBinary=function(e,t){return r(this,2,e,t)},M.toDecimalPlaces=M.toDP=function(e,t){var r=this,n=r.constructor;return r=new n(r),void 0===e?r:(A(e,0,d),void 0===t?t=n.rounding:A(t,0,8),U(r,e+r.e+1,t))},M.toExponential=function(e,t){var r,n=this,i=n.constructor;return void 0===e?r=_(n,!0):(A(e,0,d),void 0===t?t=i.rounding:A(t,0,8),r=_(n=U(new i(n),e+1,t),!0,e+1)),n.isNeg()&&!n.isZero()?"-"+r:r},M.toFixed=function(e,t){var r,n,i=this.constructor;return void 0===e?r=_(this):(A(e,0,d),void 0===t?t=i.rounding:A(t,0,8),r=_(n=U(new i(this),e+this.e+1,t),!1,e+n.e+1)),this.isNeg()&&!this.isZero()?"-"+r:r},M.toFraction=function(e){var t,r,n,i,o,a,s,u,c,f,l,p,m=this.d,h=this.constructor;if(!m)return new h(this);if(c=r=new h(1),a=(o=(t=new h(n=u=new h(0))).e=q(m)-this.e-1)%R,t.d[0]=g(10,a<0?R+a:a),null==e)e=0<o?t:c;else{if(!(s=new h(e)).isInt()||s.lt(c))throw Error(y+s);e=s.gt(t)?0<o?t:c:s}for(x=!1,s=new h(E(m)),f=h.precision,h.precision=o=m.length*R*2;l=T(s,t,0,1,1),1!=(i=r.plus(l.times(n))).cmp(e);)r=n,n=i,i=c,c=u.plus(l.times(i)),u=i,i=t,t=s.minus(l.times(i)),s=i;return i=T(e.minus(r),n,0,1,1),u=u.plus(i.times(c)),r=r.plus(i.times(n)),u.s=c.s=this.s,p=T(c,n,o,1).minus(this).abs().cmp(T(u,r,o,1).minus(this).abs())<1?[c,n]:[u,r],h.precision=f,x=!0,p},M.toHexadecimal=M.toHex=function(e,t){return r(this,16,e,t)},M.toNearest=function(e,t){var r=this,n=r.constructor;if(r=new n(r),null==e){if(!r.d)return r;e=new n(1),t=n.rounding}else{if(e=new n(e),void 0===t?t=n.rounding:A(t,0,8),!r.d)return e.s?r:e;if(!e.d)return e.s&&(e.s=r.s),e}return e.d[0]?(x=!1,r=T(r,e,0,t,1).times(e),x=!0,U(r)):(e.s=r.s,r=e),r},M.toNumber=function(){return+this},M.toOctal=function(e,t){return r(this,8,e,t)},M.toPower=M.pow=function(e){var t,r,n,i,o,a,s=this,u=s.constructor,c=+(e=new u(e));if(!(s.d&&e.d&&s.d[0]&&e.d[0]))return new u(g(+s,c));if((s=new u(s)).eq(1))return s;if(n=u.precision,o=u.rounding,e.eq(1))return U(s,n,o);if((t=k(e.e/R))>=e.d.length-1&&(r=c<0?-c:c)<=9007199254740991)return i=L(u,s,r,n),e.s<0?new u(1).div(i):U(i,n,o);if((a=s.s)<0){if(t<e.d.length-1)return new u(NaN);if(0==(1&e.d[t])&&(a=1),0==s.e&&1==s.d[0]&&1==s.d.length)return s.s=a,s}return(t=0!=(r=g(+s,c))&&isFinite(r)?new u(r+"").e:k(c*(Math.log("0."+E(s.d))/Math.LN10+s.e+1)))>u.maxE+1||t<u.minE-1?new u(0<t?a/0:0):(x=!1,u.rounding=s.s=1,r=Math.min(12,(t+"").length),(i=j(e.times($(s,n+r)),n)).d&&S((i=U(i,n+5,1)).d,n,o)&&(t=n+10,+E((i=U(j(e.times($(s,t+r)),t),t+5,1)).d).slice(n+1,n+15)+1==1e14&&(i=U(i,n+1,0))),i.s=a,x=!0,U(i,n,u.rounding=o))},M.toPrecision=function(e,t){var r,n=this,i=n.constructor;return void 0===e?r=_(n,n.e<=i.toExpNeg||n.e>=i.toExpPos):(A(e,1,d),void 0===t?t=i.rounding:A(t,0,8),r=_(n=U(new i(n),e,t),e<=n.e||n.e<=i.toExpNeg,e)),n.isNeg()&&!n.isZero()?"-"+r:r},M.toSignificantDigits=M.toSD=function(e,t){var r=this.constructor;return void 0===e?(e=r.precision,t=r.rounding):(A(e,1,d),void 0===t?t=r.rounding:A(t,0,8)),U(new r(this),e,t)},M.toString=function(){var e=this.constructor,t=_(this,this.e<=e.toExpNeg||this.e>=e.toExpPos);return this.isNeg()&&!this.isZero()?"-"+t:t},M.truncated=M.trunc=function(){return U(new this.constructor(this),this.e+1,1)},M.valueOf=M.toJSON=function(){var e=this.constructor,t=_(this,this.e<=e.toExpNeg||this.e>=e.toExpPos);return this.isNeg()?"-"+t:t};var T=function(){function C(e,t,r){var n,i=0,o=e.length;for(e=e.slice();o--;)n=e[o]*t+i,e[o]=n%r|0,i=n/r|0;return i&&e.unshift(i),e}function z(e,t,r,n){var i,o;if(r!=n)o=n<r?1:-1;else for(i=o=0;i<r;i++)if(e[i]!=t[i]){o=e[i]>t[i]?1:-1;break}return o}function B(e,t,r,n){for(var i=0;r--;)e[r]-=i,i=e[r]<t[r]?1:0,e[r]=i*n+e[r]-t[r];for(;!e[0]&&1<e.length;)e.shift()}return function(e,t,r,n,i,o){var a,s,u,c,f,l,p,m,h,d,v,y,g,x,b,w,N,M,E,A,S=e.constructor,O=e.s==t.s?1:-1,T=e.d,_=t.d;if(!(T&&T[0]&&_&&_[0]))return new S(e.s&&t.s&&(T?!_||T[0]!=_[0]:_)?T&&0==T[0]||!_?0*O:O/0:NaN);for(o?(f=1,s=e.e-t.e):(o=P,f=R,s=k(e.e/f)-k(t.e/f)),E=_.length,N=T.length,d=(h=new S(O)).d=[],u=0;_[u]==(T[u]||0);u++);if(_[u]>(T[u]||0)&&s--,null==r?(x=r=S.precision,n=S.rounding):x=i?r+(e.e-t.e)+1:r,x<0)d.push(1),l=!0;else{if(x=x/f+2|0,u=0,1==E){for(_=_[c=0],x++;(u<N||c)&&x--;u++)b=c*o+(T[u]||0),d[u]=b/_|0,c=b%_|0;l=c||u<N}else{for(1<(c=o/(_[0]+1)|0)&&(_=C(_,c,o),T=C(T,c,o),E=_.length,N=T.length),w=E,y=(v=T.slice(0,E)).length;y<E;)v[y++]=0;for((A=_.slice()).unshift(0),M=_[0],_[1]>=o/2&&++M;c=0,(a=z(_,v,E,y))<0?(g=v[0],E!=y&&(g=g*o+(v[1]||0)),1<(c=g/M|0)?(o<=c&&(c=o-1),1==(a=z(p=C(_,c,o),v,m=p.length,y=v.length))&&(c--,B(p,E<m?A:_,m,o))):(0==c&&(a=c=1),p=_.slice()),(m=p.length)<y&&p.unshift(0),B(v,p,y,o),-1==a&&(a=z(_,v,E,y=v.length))<1&&(c++,B(v,E<y?A:_,y,o)),y=v.length):0===a&&(c++,v=[0]),d[u++]=c,a&&v[0]?v[y++]=T[w]||0:(v=[T[w]],y=1),(w++<N||void 0!==v[0])&&x--;);l=void 0!==v[0]}d[0]||d.shift()}if(1==f)h.e=s,I=l;else{for(u=1,c=d[0];10<=c;c/=10)u++;h.e=u+s*f-1,U(h,i?r+h.e+1:r,n,l)}return h}}();function U(e,t,r,n){var i,o,a,s,u,c,f,l,p,m=e.constructor;e:if(null!=t){if(!(l=e.d))return e;for(i=1,s=l[0];10<=s;s/=10)i++;if((o=t-i)<0)o+=R,a=t,u=(f=l[p=0])/g(10,i-a-1)%10|0;else if(p=Math.ceil((o+1)/R),(s=l.length)<=p){if(!n)break e;for(;s++<=p;)l.push(0);f=u=0,a=(o%=R)-R+(i=1)}else{for(f=s=l[p],i=1;10<=s;s/=10)i++;u=(a=(o%=R)-R+i)<0?0:f/g(10,i-a-1)%10|0}if(n=n||t<0||void 0!==l[p+1]||(a<0?f:f%g(10,i-a-1)),c=r<4?(u||n)&&(0==r||r==(e.s<0?3:2)):5<u||5==u&&(4==r||n||6==r&&(0<o?0<a?f/g(10,i-a):0:l[p-1])%10&1||r==(e.s<0?8:7)),t<1||!l[0])return l.length=0,c?(t-=e.e+1,l[0]=g(10,(R-t%R)%R),e.e=-t||0):l[0]=e.e=0,e;if(0==o?(l.length=p,s=1,p--):(l.length=p+1,s=g(10,R-o),l[p]=0<a?(f/g(10,i-a)%g(10,a)|0)*s:0),c)for(;;){if(0==p){for(o=1,a=l[0];10<=a;a/=10)o++;for(a=l[0]+=s,s=1;10<=a;a/=10)s++;o!=s&&(e.e++,l[0]==P&&(l[0]=1));break}if(l[p]+=s,l[p]!=P)break;l[p--]=0,s=1}for(o=l.length;0===l[--o];)l.pop()}return x&&(e.e>m.maxE?(e.d=null,e.e=NaN):e.e<m.minE&&(e.e=0,e.d=[0])),e}function _(e,t,r){if(!e.isFinite())return H(e);var n,i=e.e,o=E(e.d),a=o.length;return t?(r&&0<(n=r-a)?o=o.charAt(0)+"."+o.slice(1)+D(n):1<a&&(o=o.charAt(0)+"."+o.slice(1)),o=o+(e.e<0?"e":"e+")+e.e):i<0?(o="0."+D(-i-1)+o,r&&0<(n=r-a)&&(o+=D(n))):a<=i?(o+=D(i+1-a),r&&0<(n=r-i-1)&&(o=o+"."+D(n))):((n=i+1)<a&&(o=o.slice(0,n)+"."+o.slice(n)),r&&0<(n=r-a)&&(i+1===a&&(o+="."),o+=D(n))),o}function C(e,t){var r=e[0];for(t*=R;10<=r;r/=10)t++;return t}function z(e,t,r){if(w<t)throw x=!0,r&&(e.precision=r),Error(o);return U(new e(n),t,1,!0)}function B(e,t,r){if(N<t)throw Error(o);return U(new e(i),t,r,!0)}function q(e){var t=e.length-1,r=t*R+1;if(t=e[t]){for(;t%10==0;t/=10)r--;for(t=e[0];10<=t;t/=10)r++}return r}function D(e){for(var t="";e--;)t+="0";return t}function L(e,t,r,n){var i,o=new e(1),a=Math.ceil(n/R+4);for(x=!1;;){if(r%2&&W((o=o.times(t)).d,a)&&(i=!0),0===(r=k(r/2))){r=o.d.length-1,i&&0===o.d[r]&&++o.d[r];break}W((t=t.times(t)).d,a)}return x=!0,o}function F(e){return 1&e.d[e.d.length-1]}function t(e,t,r){for(var n,i=new e(t[0]),o=0;++o<t.length;){if(!(n=new e(t[o])).s){i=n;break}i[r](n)&&(i=n)}return i}function j(e,t){var r,n,i,o,a,s,u,c=0,f=0,l=0,p=e.constructor,m=p.rounding,h=p.precision;if(!e.d||!e.d[0]||17<e.e)return new p(e.d?e.d[0]?e.s<0?0:1/0:1:e.s?e.s<0?0:e:NaN);for(null==t?(x=!1,u=h):u=t,s=new p(.03125);-2<e.e;)e=e.times(s),l+=5;for(u+=n=Math.log(g(2,l))/Math.LN10*2+5|0,r=o=a=new p(1),p.precision=u;;){if(o=U(o.times(e),u,1),r=r.times(++f),E((s=a.plus(T(o,r,u,1))).d).slice(0,u)===E(a.d).slice(0,u)){for(i=l;i--;)a=U(a.times(a),u,1);if(null!=t)return p.precision=h,a;if(!(c<3&&S(a.d,u-n,m,c)))return U(a,p.precision=h,m,x=!0);p.precision=u+=10,r=o=s=new p(1),f=0,c++}a=s}}function $(e,t){var r,n,i,o,a,s,u,c,f,l,p,m=1,h=e,d=h.d,v=h.constructor,y=v.rounding,g=v.precision;if(h.s<0||!d||!d[0]||!h.e&&1==d[0]&&1==d.length)return new v(d&&!d[0]?-1/0:1!=h.s?NaN:d?0:h);if(null==t?(x=!1,f=g):f=t,v.precision=f+=10,n=(r=E(d)).charAt(0),!(Math.abs(o=h.e)<15e14))return c=z(v,f+2,g).times(o+""),h=$(new v(n+"."+r.slice(1)),f-10).plus(c),v.precision=g,null==t?U(h,g,y,x=!0):h;for(;n<7&&1!=n||1==n&&3<r.charAt(1);)n=(r=E((h=h.times(e)).d)).charAt(0),m++;for(o=h.e,1<n?(h=new v("0."+r),o++):h=new v(n+"."+r.slice(1)),u=a=h=T((l=h).minus(1),h.plus(1),f,1),p=U(h.times(h),f,1),i=3;;){if(a=U(a.times(p),f,1),E((c=u.plus(T(a,new v(i),f,1))).d).slice(0,f)===E(u.d).slice(0,f)){if(u=u.times(2),0!==o&&(u=u.plus(z(v,f+2,g).times(o+""))),u=T(u,new v(m),f,1),null!=t)return v.precision=g,u;if(!S(u.d,f-10,y,s))return U(u,v.precision=g,y,x=!0);v.precision=f+=10,c=a=h=T(l.minus(1),l.plus(1),f,1),p=U(h.times(h),f,1),i=s=1}u=c,i+=2}}function H(e){return String(e.s*e.s/0)}function G(e,t){var r,n,i;for(-1<(r=t.indexOf("."))&&(t=t.replace(".","")),0<(n=t.search(/e/i))?(r<0&&(r=n),r+=+t.slice(n+1),t=t.substring(0,n)):r<0&&(r=t.length),n=0;48===t.charCodeAt(n);n++);for(i=t.length;48===t.charCodeAt(i-1);--i);if(t=t.slice(n,i)){if(i-=n,e.e=r=r-n-1,e.d=[],n=(r+1)%R,r<0&&(n+=R),n<i){for(n&&e.d.push(+t.slice(0,n)),i-=R;n<i;)e.d.push(+t.slice(n,n+=R));t=t.slice(n),n=R-t.length}else n-=i;for(;n--;)t+="0";e.d.push(+t),x&&(e.e>e.constructor.maxE?(e.d=null,e.e=NaN):e.e<e.constructor.minE&&(e.e=0,e.d=[0]))}else e.e=0,e.d=[0];return e}function Z(e,t,r,n,i){var o,a,s,u,c=e.precision,f=Math.ceil(c/R);for(x=!1,u=r.times(r),s=new e(n);;){if(a=T(s.times(u),new e(t++*t++),c,1),s=i?n.plus(a):n.minus(a),n=T(a.times(u),new e(t++*t++),c,1),void 0!==(a=s.plus(n)).d[f]){for(o=f;a.d[o]===s.d[o]&&o--;);if(-1==o)break}o=s,s=n,n=a,a=o,0}return x=!0,a.d.length=f+1,a}function V(e,t){var r,n=t.s<0,i=B(e,e.precision,1),o=i.times(.5);if((t=t.abs()).lte(o))return a=n?4:1,t;if((r=t.divToInt(i)).isZero())a=n?3:2;else{if((t=t.minus(r.times(i))).lte(o))return a=F(r)?n?2:3:n?4:1,t;a=F(r)?n?1:4:n?3:2}return t.minus(i).abs()}function r(e,t,r,n){var i,o,a,s,u,c,f,l,p,m=e.constructor,h=void 0!==r;if(h?(A(r,1,d),void 0===n?n=m.rounding:A(n,0,8)):(r=m.precision,n=m.rounding),e.isFinite()){for(h?(i=2,16==t?r=4*r-3:8==t&&(r=3*r-2)):i=t,0<=(a=(f=_(e)).indexOf("."))&&(f=f.replace(".",""),(p=new m(1)).e=f.length-a,p.d=O(_(p),10,i),p.e=p.d.length),o=u=(l=O(f,10,i)).length;0==l[--u];)l.pop();if(l[0]){if(a<0?o--:((e=new m(e)).d=l,e.e=o,l=(e=T(e,p,r,n,0,i)).d,o=e.e,c=I),a=l[r],s=i/2,c=c||void 0!==l[r+1],c=n<4?(void 0!==a||c)&&(0===n||n===(e.s<0?3:2)):s<a||a===s&&(4===n||c||6===n&&1&l[r-1]||n===(e.s<0?8:7)),l.length=r,c)for(;++l[--r]>i-1;)l[r]=0,r||(++o,l.unshift(1));for(u=l.length;!l[u-1];--u);for(a=0,f="";a<u;a++)f+=v.charAt(l[a]);if(h){if(1<u)if(16==t||8==t){for(a=16==t?4:3,--u;u%a;u++)f+="0";for(u=(l=O(f,i,t)).length;!l[u-1];--u);for(a=1,f="1.";a<u;a++)f+=v.charAt(l[a])}else f=f.charAt(0)+"."+f.slice(1);f=f+(o<0?"p":"p+")+o}else if(o<0){for(;++o;)f="0"+f;f="0."+f}else if(++o>u)for(o-=u;o--;)f+="0";else o<u&&(f=f.slice(0,o)+"."+f.slice(o))}else f=h?"0p+0":"0";f=(16==t?"0x":2==t?"0b":8==t?"0o":"")+f}else f=H(e);return e.s<0?"-"+f:f}function W(e,t){if(e.length>t)return e.length=t,!0}function J(e){return new this(e).abs()}function Y(e){return new this(e).acos()}function X(e){return new this(e).acosh()}function Q(e,t){return new this(e).plus(t)}function K(e){return new this(e).asin()}function ee(e){return new this(e).asinh()}function te(e){return new this(e).atan()}function re(e){return new this(e).atanh()}function ne(e,t){e=new this(e),t=new this(t);var r,n=this.precision,i=this.rounding,o=n+4;return e.s&&t.s?e.d||t.d?!t.d||e.isZero()?(r=t.s<0?B(this,n,i):new this(0)).s=e.s:!e.d||t.isZero()?(r=B(this,o,1).times(.5)).s=e.s:t.s<0?(this.precision=o,this.rounding=1,r=this.atan(T(e,t,o,1)),t=B(this,o,1),this.precision=n,this.rounding=i,r=e.s<0?r.minus(t):r.plus(t)):r=this.atan(T(e,t,o,1)):(r=B(this,o,1).times(0<t.s?.25:.75)).s=e.s:r=new this(NaN),r}function ie(e){return new this(e).cbrt()}function oe(e){return U(e=new this(e),e.e+1,2)}function ae(e){if(!e||"object"!=typeof e)throw Error(c+"Object expected");var t,r,n,i=!0===e.defaults,o=["precision",1,d,"rounding",0,8,"toExpNeg",-s,0,"toExpPos",0,s,"maxE",0,s,"minE",-s,0,"modulo",0,9];for(t=0;t<o.length;t+=3)if(r=o[t],i&&(this[r]=u[r]),void 0!==(n=e[r])){if(!(k(n)===n&&o[t+1]<=n&&n<=o[t+2]))throw Error(y+r+": "+n);this[r]=n}if(r="crypto",i&&(this[r]=u[r]),void 0!==(n=e[r])){if(!0!==n&&!1!==n&&0!==n&&1!==n)throw Error(y+r+": "+n);if(n){if("undefined"==typeof crypto||!crypto||!crypto.getRandomValues&&!crypto.randomBytes)throw Error(f);this[r]=!0}else this[r]=!1}return this}function se(e){return new this(e).cos()}function ue(e){return new this(e).cosh()}function ce(e,t){return new this(e).div(t)}function fe(e){return new this(e).exp()}function le(e){return U(e=new this(e),e.e+1,3)}function pe(){var e,t,r=new this(0);for(x=!1,e=0;e<arguments.length;)if((t=new this(arguments[e++])).d)r.d&&(r=r.plus(t.times(t)));else{if(t.s)return x=!0,new this(1/0);r=t}return x=!0,r.sqrt()}function me(e){return e instanceof l||e&&"[object Decimal]"===e.name||!1}function he(e){return new this(e).ln()}function de(e,t){return new this(e).log(t)}function ve(e){return new this(e).log(2)}function ye(e){return new this(e).log(10)}function ge(){return t(this,arguments,"lt")}function xe(){return t(this,arguments,"gt")}function be(e,t){return new this(e).mod(t)}function we(e,t){return new this(e).mul(t)}function Ne(e,t){return new this(e).pow(t)}function Me(e){var t,r,n,i,o=0,a=new this(1),s=[];if(void 0===e?e=this.precision:A(e,1,d),n=Math.ceil(e/R),this.crypto)if(crypto.getRandomValues)for(t=crypto.getRandomValues(new Uint32Array(n));o<n;)429e7<=(i=t[o])?t[o]=crypto.getRandomValues(new Uint32Array(1))[0]:s[o++]=i%1e7;else{if(!crypto.randomBytes)throw Error(f);for(t=crypto.randomBytes(n*=4);o<n;)214e7<=(i=t[o]+(t[o+1]<<8)+(t[o+2]<<16)+((127&t[o+3])<<24))?crypto.randomBytes(4).copy(t,o):(s.push(i%1e7),o+=4);o=n/4}else for(;o<n;)s[o++]=1e7*Math.random()|0;for(n=s[--o],e%=R,n&&e&&(i=g(10,R-e),s[o]=(n/i|0)*i);0===s[o];o--)s.pop();if(o<0)s=[r=0];else{for(r=-1;0===s[0];r-=R)s.shift();for(n=1,i=s[0];10<=i;i/=10)n++;n<R&&(r-=R-n)}return a.e=r,a.d=s,a}function Ee(e){return U(e=new this(e),e.e+1,this.rounding)}function Ae(e){return(e=new this(e)).d?e.d[0]?e.s:0*e.s:e.s||NaN}function Se(e){return new this(e).sin()}function Oe(e){return new this(e).sinh()}function Te(e){return new this(e).sqrt()}function _e(e,t){return new this(e).sub(t)}function Ce(e){return new this(e).tan()}function ze(e){return new this(e).tanh()}function Be(e){return U(e=new this(e),e.e+1,1)}(l=function e(t){var r,n,i;function o(e){var t,r,n,i=this;if(!(i instanceof o))return new o(e);if(e instanceof(i.constructor=o))return i.s=e.s,i.e=e.e,void(i.d=(e=e.d)?e.slice():e);if("number"==(n=typeof e)){if(0===e)return i.s=1/e<0?-1:1,i.e=0,void(i.d=[0]);if(e<0?(e=-e,i.s=-1):i.s=1,e===~~e&&e<1e7){for(t=0,r=e;10<=r;r/=10)t++;return i.e=t,void(i.d=[e])}return 0*e!=0?(e||(i.s=NaN),i.e=NaN,void(i.d=null)):G(i,e.toString())}if("string"!==n)throw Error(y+e);return 45===e.charCodeAt(0)?(e=e.slice(1),i.s=-1):i.s=1,b.test(e)?G(i,e):function(e,t){var r,n,i,o,a,s,u,c,f;if("Infinity"===t||"NaN"===t)return+t||(e.s=NaN),e.e=NaN,e.d=null,e;if(m.test(t))r=16,t=t.toLowerCase();else if(p.test(t))r=2;else{if(!h.test(t))throw Error(y+t);r=8}for(0<(o=t.search(/p/i))?(u=+t.slice(o+1),t=t.substring(2,o)):t=t.slice(2),a=0<=(o=t.indexOf(".")),n=e.constructor,a&&(o=(s=(t=t.replace(".","")).length)-o,i=L(n,new n(r),o,2*o)),o=f=(c=O(t,r,P)).length-1;0===c[o];--o)c.pop();return o<0?new n(0*e.s):(e.e=C(c,f),e.d=c,x=!1,a&&(e=T(e,i,4*s)),u&&(e=e.times(Math.abs(u)<54?Math.pow(2,u):l.pow(2,u))),x=!0,e)}(i,e)}if(o.prototype=M,o.ROUND_UP=0,o.ROUND_DOWN=1,o.ROUND_CEIL=2,o.ROUND_FLOOR=3,o.ROUND_HALF_UP=4,o.ROUND_HALF_DOWN=5,o.ROUND_HALF_EVEN=6,o.ROUND_HALF_CEIL=7,o.ROUND_HALF_FLOOR=8,o.EUCLID=9,o.config=o.set=ae,o.clone=e,o.isDecimal=me,o.abs=J,o.acos=Y,o.acosh=X,o.add=Q,o.asin=K,o.asinh=ee,o.atan=te,o.atanh=re,o.atan2=ne,o.cbrt=ie,o.ceil=oe,o.cos=se,o.cosh=ue,o.div=ce,o.exp=fe,o.floor=le,o.hypot=pe,o.ln=he,o.log=de,o.log10=ye,o.log2=ve,o.max=ge,o.min=xe,o.mod=be,o.mul=we,o.pow=Ne,o.random=Me,o.round=Ee,o.sign=Ae,o.sin=Se,o.sinh=Oe,o.sqrt=Te,o.sub=_e,o.tan=Ce,o.tanh=ze,o.trunc=Be,void 0===t&&(t={}),t&&!0!==t.defaults)for(i=["precision","rounding","toExpNeg","toExpPos","maxE","minE","modulo","crypto"],r=0;r<i.length;)t.hasOwnProperty(n=i[r++])||(t[n]=this[n]);return o.config(t),o}(u)).default=l.Decimal=l,n=new l(n),i=new l(i),void 0===(Re=function(){return l}.call(ke,Pe,ke,Ie))||(Ie.exports=Re)}()},function(e,t,r){"use strict";var n,i=r(559),a=(n=i)&&n.__esModule?n:{default:n};t.name="BigNumber",t.path="type",t.factory=function(e,t,r,n,i){var o=a.default.clone({precision:t.precision});return o.prototype.type="BigNumber",o.prototype.isBigNumber=!0,o.prototype.toJSON=function(){return{mathjs:"BigNumber",value:this.toString()}},o.fromJSON=function(e){return new o(e.value)},i.on("config",function(e,t){e.precision!==t.precision&&o.config({precision:e.precision})}),o},t.math=!0},function(e,t,r){"use strict";e.exports=[r(560),r(558)]},function(e,t,r){"use strict";e.exports=[r(561),r(557),r(556),r(552),r(548),r(545),r(88),r(537),r(536),r(535)]},function(e,t,r){"use strict";e.exports=[r(562),r(529),r(527),r(371),r(151),r(149)]},function(e,t,r){"use strict";var u=r(5);function c(e,t,r){if(void 0!==e[t]&&(a=r,s=e[t],-1===a.indexOf(s))){var n=(i=r,o=e[t],i.map(function(e){return e.toLowerCase()}).indexOf(o.toLowerCase()));-1!==n?(console.warn('Warning: Wrong casing for configuration option "'+t+'", should be "'+r[n]+'" instead of "'+e[t]+'".'),e[t]=r[n]):console.warn('Warning: Unknown value "'+e[t]+'" for configuration option "'+t+'". Available options: '+r.map(JSON.stringify).join(", ")+".")}var i,o,a,s}t.name="config",t.math=!0,t.factory=function(e,i,t,r,o){var a=["Matrix","Array"],s=["number","BigNumber","Fraction"];function n(e){if(e){var t=u.map(i,u.clone);c(e,"matrix",a),c(e,"number",s),u.deepExtend(i,e);var r=u.map(i,u.clone),n=u.map(e,u.clone);return o.emit("config",r,t,n),r}return u.map(i,u.clone)}return n.MATRIX=a,n.NUMBER=s,n}},function(e,t,r){"use strict";var v="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e},y=r(5).lazy,g=r(5).isFactory,x=r(5).traverse,b=r(59);t.math=!0,t.name="import",t.factory=function(s,e,u,c,f){function l(e,t,r){if(r.wrap&&"function"==typeof t&&(n=function(){for(var e=[],t=0,r=arguments.length;t<r;t++){var n=arguments[t];e[t]=n&&n.valueOf()}return i.apply(f,e)},(i=t).transform&&(n.transform=i.transform),t=n),h(f[e])&&h(t))return t=r.override?c(e,t.signatures):c(f[e],t),o(e,f[e]=t),void f.emit("import",e,function(){return t});var i,n;if(void 0===f[e]||r.override)return o(e,f[e]=t),void f.emit("import",e,function(){return t});if(!r.silent)throw new Error('Cannot import "'+e+'": already exists')}function o(e,t){t&&"function"==typeof t.transform?(f.expression.transform[e]=t.transform,r(e)&&(f.expression.mathWithTransform[e]=t.transform)):(delete f.expression.transform[e],r(e)&&(f.expression.mathWithTransform[e]=t))}function p(e){delete f.expression.transform[e],r(e)?f.expression.mathWithTransform[e]=f[e]:delete f.expression.mathWithTransform[e]}function m(t,r){if("string"==typeof t.name){var n=t.name,e=n in f.expression.transform,i=t.path?x(f,t.path):f,o=i.hasOwnProperty(n)?i[n]:void 0,a=function(){var e=u(t);if(e&&"function"==typeof e.transform)throw new Error('Transforms cannot be attached to factory functions. Please create a separate function for it with exports.path="expression.transform"');if(h(o)&&h(e))return r.override||(e=c(o,e)),e;if(void 0===o||r.override)return e;if(!r.silent)throw new Error('Cannot import "'+n+'": already exists')};!1!==t.lazy?(y(i,n,a),e?p(n):("expression.transform"===t.path||d(t))&&y(f.expression.mathWithTransform,n,a)):(i[n]=a(),e?p(n):("expression.transform"===t.path||d(t))&&(f.expression.mathWithTransform[n]=a())),f.emit("import",n,a,t.path)}else u(t)}function h(e){return"function"==typeof e&&"object"===v(e.signatures)}function r(e){return!t.hasOwnProperty(e)}function d(e){return void 0===e.path&&!t.hasOwnProperty(e.name)}var t={expression:!0,type:!0,docs:!0,error:!0,json:!0,chain:!0};return function t(e,r){var n,i=arguments.length;if(1!==i&&2!==i)throw new b("import",i,1,2);if(r||(r={}),g(e))m(e,r);else if(Array.isArray(e))e.forEach(function(e){t(e,r)});else if("object"===(void 0===e?"undefined":v(e))){for(var o in e)if(e.hasOwnProperty(o)){var a=e[o];"function"==typeof(n=a)||"number"==typeof n||"string"==typeof n||"boolean"==typeof n||null===n||n&&s.isUnit(n)||n&&s.isComplex(n)||n&&s.isBigNumber(n)||n&&s.isFraction(n)||n&&s.isMatrix(n)||n&&Array.isArray(n)?l(o,a,r):g(e)?m(e,r):t(a,r)}}else if(!r.silent)throw new TypeError("Factory, Object, or Array expected")}},t.lazy=!0},function(e,t){function r(){}r.prototype={on:function(e,t,r){var n=this.e||(this.e={});return(n[e]||(n[e]=[])).push({fn:t,ctx:r}),this},once:function(e,t,r){var n=this;function i(){n.off(e,i),t.apply(r,arguments)}return i._=t,this.on(e,i,r)},emit:function(e){for(var t=[].slice.call(arguments,1),r=((this.e||(this.e={}))[e]||[]).slice(),n=0,i=r.length;n<i;n++)r[n].fn.apply(r[n].ctx,t);return this},off:function(e,t){var r=this.e||(this.e={}),n=r[e],i=[];if(n&&t)for(var o=0,a=n.length;o<a;o++)n[o].fn!==t&&n[o].fn._!==t&&i.push(n[o]);return i.length?r[e]=i:delete r[e],this}},e.exports=r},function(e,t,r){"use strict";var n,i,o;i=[],void 0===(o="function"==typeof(n=function(){function y(){return!0}function ce(){return!1}function fe(){}return function e(){var t=[{name:"number",test:function(e){return"number"==typeof e}},{name:"string",test:function(e){return"string"==typeof e}},{name:"boolean",test:function(e){return"boolean"==typeof e}},{name:"Function",test:function(e){return"function"==typeof e}},{name:"Array",test:Array.isArray},{name:"Date",test:function(e){return e instanceof Date}},{name:"RegExp",test:function(e){return e instanceof RegExp}},{name:"Object",test:function(e){return"object"==typeof e&&e.constructor===Object}},{name:"null",test:function(e){return null===e}},{name:"undefined",test:function(e){return void 0===e}}],n={name:"any",test:y},r=[],i=[],j={types:t,conversions:i,ignore:r};function f(t){var e=se(j.types,function(e){return e.name===t});if(e)return e;if("any"===t)return n;var r=se(j.types,function(e){return e.name.toLowerCase()===t.toLowerCase()});throw new TypeError('Unknown type "'+t+'"'+(r?'. Did you mean "'+r.name+'"?':""))}function l(e){return e===n?999:j.types.indexOf(e)}function $(t){var e=se(j.types,function(e){return e.test(t)});if(e)return e.name;throw new TypeError("Value has unknown type. Value: "+t)}function H(e){return e.map(function(e){var t=e.types.map(o);return(e.restParam?"...":"")+t.join("|")}).join(",")}function G(e,r){var t,n,i=0===e.indexOf("..."),o=i?3<e.length?e.slice(3):"any":e,a=o.split("|").map(re).filter(h).filter(m),s=(t=a,n={},r.forEach(function(e){-1!==t.indexOf(e.from)||-1===t.indexOf(e.to)||n[e.from]||(n[e.from]=e)}),Object.keys(n).map(function(e){return n[e]})),u=a.map(function(e){var t=f(e);return{name:e,typeIndex:l(t),test:t.test,conversion:null,conversionIndex:-1}}),c=s.map(function(e){var t=f(e.from);return{name:e.from,typeIndex:l(t),test:t.test,conversion:e,conversionIndex:r.indexOf(e)}});return{types:u.concat(c),restParam:i}}function Z(e){var t=d(e);return!!t&&t.restParam}function V(e){return e.types.some(function(e){return null!=e.conversion})}function W(e){if(e&&0!==e.types.length){if(1===e.types.length)return f(e.types[0].name).test;if(2===e.types.length){var t=f(e.types[0].name).test,r=f(e.types[1].name).test;return function(e){return t(e)||r(e)}}var n=e.types.map(function(e){return f(e.name).test});return function(e){for(var t=0;t<n.length;t++)if(n[t](e))return!0;return!1}}return y}function J(e){var r,t,n,i;if(Z(e)){var o=(r=(i=e,i.slice(0,i.length-1)).map(W)).length,a=W(d(e));return function(e){for(var t=0;t<r.length;t++)if(!r[t](e[t]))return!1;return function(e){for(var t=o;t<e.length;t++)if(!a(e[t]))return!1;return!0}(e)&&e.length>=o+1}}return 0===e.length?function(e){return 0===e.length}:1===e.length?(t=W(e[0]),function(e){return t(e[0])&&1===e.length}):2===e.length?(t=W(e[0]),n=W(e[1]),function(e){return t(e[0])&&n(e[1])&&2===e.length}):(r=e.map(W),function(e){for(var t=0;t<r.length;t++)if(!r[t](e[t]))return!1;return e.length===r.length})}function Y(e,t){return t<e.params.length?e.params[t]:Z(e.params)?d(e.params):null}function X(e,t,r){var n=Y(e,t),i=n?r?n.types.filter(p):n.types:[];return i.map(o)}function o(e){return e.name}function p(e){return null===e.conversion||void 0===e.conversion}function Q(e,t){var r=function(e){for(var t={},r=0;r<e.length;r++)t[e[r]]=!0;return Object.keys(t)}(ue(e,function(e){return X(e,t,!1)}));return-1!==r.indexOf("any")?["any"]:r}function a(e){for(var t=999,r=0;r<e.types.length;r++)p(e.types[r])&&(t=Math.min(t,e.types[r].typeIndex));return t}function s(e){for(var t=999,r=0;r<e.types.length;r++)p(e.types[r])||(t=Math.min(t,e.types[r].conversionIndex));return t}function u(e,t){var r;return 0!=(r=e.restParam-t.restParam)?r:0!=(r=V(e)-V(t))?r:0!=(r=a(e)-a(t))?r:s(e)-s(t)}function K(e,t){var r,n,i=Math.min(e.params.length,t.params.length);if(0!=(n=e.params.some(V)-t.params.some(V)))return n;for(r=0;r<i;r++)if(0!=(n=V(e.params[r])-V(t.params[r])))return n;for(r=0;r<i;r++)if(0!==(n=u(e.params[r],t.params[r])))return n;return e.params.length-t.params.length}function ee(e){var t,r,n,i,o=[],a=[];switch(e.types.forEach(function(e){e.conversion&&(o.push(f(e.conversion.from).test),a.push(e.conversion.convert))}),a.length){case 0:return function(e){return e};case 1:return t=o[0],n=a[0],function(e){return t(e)?n(e):e};case 2:return t=o[0],r=o[1],n=a[0],i=a[1],function(e){return t(e)?n(e):r(e)?i(e):e};default:return function(e){for(var t=0;t<a.length;t++)if(o[t](e))return a[t](e);return e}}}function te(e,c){return function t(r,n,i){if(n<r.length){var e,o=r[n],a=c?o.types.filter(p):o.types;if(o.restParam){var s=a.filter(p);e=s.length<a.length?[s,a]:[a]}else e=a.map(function(e){return[e]});return ue(e,function(e){return t(r,n+1,i.concat([e]))})}var u=i.map(function(e,t){return{types:e,restParam:t===r.length-1&&Z(r)}});return[u]}(e,0,[])}function c(t,o){if(0===Object.keys(o).length)throw new SyntaxError("No signatures provided");var r=[];Object.keys(o).map(function(e){return r=o[t=e],i=j.conversions,n=[],""!==t.trim()&&(n=t.split(",").map(re).map(function(e,t,r){var n=G(e,i);if(n.restParam&&t!==r.length-1)throw new SyntaxError('Unexpected rest parameter "'+e+'": only allowed for the last parameter');return n})),n.some(ie)?null:{params:n,fn:r};var t,r,i,n}).filter(ne).forEach(function(t){var e=se(r,function(e){return function(e,t){for(var r=Math.max(e.params.length,t.params.length),n=0;n<r;n++){var i=X(e,n,!0),o=X(t,n,!0);if(!ae(i,o))return!1}var a=e.params.length,s=t.params.length,u=Z(e.params),c=Z(t.params);return u?c?a===s:a<=s:c?s<=a:a===s}(e,t)});if(e)throw new TypeError('Conflicting signatures "'+H(e.params)+'" and "'+H(t.params)+'".');r.push(t)});var n=ue(r,function(t){var e=t?te(t.params,!1):[];return e.map(function(e){return{params:e,fn:t.fn}})}).filter(ne);n.sort(K);var i,e=n[0]&&n[0].params.length<=2&&!Z(n[0].params),a=n[1]&&n[1].params.length<=2&&!Z(n[1].params),s=n[2]&&n[2].params.length<=2&&!Z(n[2].params),u=n[3]&&n[3].params.length<=2&&!Z(n[3].params),c=n[4]&&n[4].params.length<=2&&!Z(n[4].params),f=n[5]&&n[5].params.length<=2&&!Z(n[5].params),l=e&&a&&s&&u&&c&&f,p=n.map(function(e){return J(e.params)}),m=e?W(n[0].params[0]):ce,h=a?W(n[1].params[0]):ce,d=s?W(n[2].params[0]):ce,v=u?W(n[3].params[0]):ce,y=c?W(n[4].params[0]):ce,g=f?W(n[5].params[0]):ce,x=e?W(n[0].params[1]):ce,b=a?W(n[1].params[1]):ce,w=s?W(n[2].params[1]):ce,N=u?W(n[3].params[1]):ce,M=c?W(n[4].params[1]):ce,E=f?W(n[5].params[1]):ce,A=n.map(function(e){return function(e,n){var t=n;if(e.some(V)){var i=Z(e),o=e.map(ee);t=function(){for(var e=[],t=i?arguments.length-1:arguments.length,r=0;r<t;r++)e[r]=o[r](arguments[r]);return i&&(e[t]=arguments[t].map(o[t])),n.apply(null,e)}}var r=t;if(Z(e)){var a=e.length-1;r=function(){return t.apply(null,oe(arguments,0,a).concat([oe(arguments,a)]))}}return r}(e.params,e.fn)}),S=e?A[0]:fe,O=a?A[1]:fe,T=s?A[2]:fe,_=u?A[3]:fe,C=c?A[4]:fe,z=f?A[5]:fe,B=e?n[0].params.length:-1,I=a?n[1].params.length:-1,k=s?n[2].params.length:-1,P=u?n[3].params.length:-1,R=c?n[4].params.length:-1,U=f?n[5].params.length:-1,q=l?6:0,D=n.length,L=function(){for(var e=q;e<D;e++)if(p[e](arguments))return A[e].apply(null,arguments);throw function(e,r,t){var n,i,o,a=e||"unnamed",s=t;for(o=0;o<r.length;o++){var u=s.filter(function(e){var t=W(Y(e,o));return(o<e.params.length||Z(e.params))&&t(r[o])});if(0===u.length){if(0<(i=Q(s,o)).length){var c=$(r[o]);return(n=new TypeError("Unexpected type of argument in function "+a+" (expected: "+i.join(" or ")+", actual: "+c+", index: "+o+")")).data={category:"wrongType",fn:a,index:o,actual:c,expected:i},n}}else s=u}var f=s.map(function(e){return Z(e.params)?1/0:e.params.length});if(r.length<Math.min.apply(null,f))return i=Q(s,o),(n=new TypeError("Too few arguments in function "+a+" (expected: "+i.join(" or ")+", index: "+r.length+")")).data={category:"tooFewArgs",fn:a,index:r.length,expected:i},n;var l=Math.max.apply(null,f);return r.length>l?(n=new TypeError("Too many arguments in function "+a+" (expected: "+l+", actual: "+r.length+")")).data={category:"tooManyArgs",fn:a,index:r.length,expectedLength:l}:(n=new TypeError('Arguments of type "'+r.join(", ")+'" do not match any of the defined signatures of function '+a+".")).data={category:"mismatch",actual:r.map($)},n}(t,arguments,n)},F=function(e,t){return arguments.length===B&&m(e)&&x(t)?S.apply(null,arguments):arguments.length===I&&h(e)&&b(t)?O.apply(null,arguments):arguments.length===k&&d(e)&&w(t)?T.apply(null,arguments):arguments.length===P&&v(e)&&N(t)?_.apply(null,arguments):arguments.length===R&&y(e)&&M(t)?C.apply(null,arguments):arguments.length===U&&g(e)&&E(t)?z.apply(null,arguments):L.apply(null,arguments)};try{Object.defineProperty(F,"name",{value:t})}catch(e){}return F.signatures=(i={},n.forEach(function(t){t.params.some(V)||te(t.params,!0).forEach(function(e){i[H(e)]=t.fn})}),i),F}function m(e){return-1===j.ignore.indexOf(e)}function re(e){return e.trim()}function h(e){return!!e}function ne(e){return null!==e}function ie(e){return 0===e.types.length}function d(e){return e[e.length-1]}function oe(e,t,r){return Array.prototype.slice.call(e,t,r)}function ae(e,t){for(var r=0;r<e.length;r++)if(n=t,i=e[r],-1!==n.indexOf(i))return!0;var n,i;return!1}function se(e,t){for(var r=0;r<e.length;r++)if(t(e[r]))return e[r]}function ue(e,t){return Array.prototype.concat.apply([],e.map(t))}function v(e){for(var t="",r=0;r<e.length;r++){var n=e[r];if(n.signatures&&""!==n.name)if(""===t)t=n.name;else if(t!==n.name){var i=new Error("Function names do not match (expected: "+t+", actual: "+n.name+")");throw i.data={actual:n.name,expected:t},i}}return t}return(j=c("typed",{"string, Object":c,Object:function(e){var t=[];for(var r in e)e.hasOwnProperty(r)&&t.push(e[r]);var n=v(t);return c(n,e)},"...Function":function(e){for(var t,r=v(e),n={},i=0;i<e.length;i++){var o=e[i];if("object"!=typeof o.signatures)throw(t=new TypeError("Function is no typed-function (index: "+i+")")).data={index:i},t;for(var a in o.signatures)if(o.signatures.hasOwnProperty(a))if(n.hasOwnProperty(a)){if(o.signatures[a]!==n[a])throw(t=new Error('Signature "'+a+'" is defined twice')).data={signature:a},t}else n[a]=o.signatures[a]}return c(r,n)}})).create=e,j.types=t,j.conversions=i,j.ignore=r,j.convert=function(e,t){var r=$(e);if(t===r)return e;for(var n=0;n<j.conversions.length;n++){var i=j.conversions[n];if(i.from===r&&i.to===t)return i.convert(e)}throw new Error("Cannot convert from "+r+" to "+t)},j.find=function(e,t){if(!e.signatures)throw new TypeError("Function is no typed-function");var r;if("string"==typeof t){r=t.split(",");for(var n=0;n<r.length;n++)r[n]=r[n].trim()}else{if(!Array.isArray(t))throw new TypeError("String array or a comma separated string expected");r=t}var i=r.join(","),o=e.signatures[i];if(o)return o;throw new TypeError("Signature not found (signature: "+(e.name||"unnamed")+"("+r.join(", ")+"))")},j.addType=function(e,t){if(!e||"string"!=typeof e.name||"function"!=typeof e.test)throw new TypeError("Object with properties {name: string, test: function} expected");if(!1!==t)for(var r=0;r<j.types.length;r++)if("Object"===j.types[r].name)return void j.types.splice(r,0,e);j.types.push(e)},j.addConversion=function(e){if(!e||"string"!=typeof e.from||"string"!=typeof e.to||"function"!=typeof e.convert)throw new TypeError("Object with properties {from: string, to: string, convert: function} expected");j.conversions.push(e)},j}()})?n.apply(t,i):n)||(e.exports=o)},function(e,t,r){"use strict";Number.isFinite=Number.isFinite||function(e){return"number"==typeof e&&isFinite(e)},Number.isNaN=Number.isNaN||function(e){return e!=e}},function(e,t,r){"use strict";r(568);var u=r(5).isFactory,c=r(148),f=r(147),l=r(565),p=r(564);t.create=function(e){if("function"!=typeof Object.create)throw new Error("ES5 not supported by this JavaScript engine. Please load the es5-shim and es5-sham library for compatibility.");var n=[],i=[],o=f.mixin({});o.type={},o.expression={transform:{},mathWithTransform:{}},o.typed=c.create(o.type);var a={epsilon:1e-12,matrix:"Matrix",number:"number",precision:64,predictable:!1,randomSeed:null};function s(e){if(!u(e))throw new Error("Factory object with properties `type`, `name`, and `factory` expected");var t=n.indexOf(e),r=void 0;return-1===t?(r=!0===e.math?e.factory(o.type,a,s,o.typed,o):e.factory(o.type,a,s,o.typed),n.push(e),i.push(r)):r=i[t],r}return o.import=s(l),o.config=s(p),o.expression.mathWithTransform.config=o.config,e&&o.config(e),o}},function(e,t,n){"use strict";var r,i=n(569),o=(r=i)&&r.__esModule?r:{default:r};e.exports=function e(t){var r=o.default.create(t);return r.create=e,r.import(n(563)),r}()}])});
-
-},{}],13:[function(require,module,exports){
-// Central tendency summary statistics
-
-var Mean = function (distribution) {
-    var size = distribution.length;
-    if (size == 0 ) {
-        return null;
-    }
-    var mean = 0; var geometricMean = 1; var rootMeanSquare = 0;
-    for (var i = 0; i < size; ++i ){
-        mean += (distribution[i]/size);
-        geometricMean *= Math.pow(distribution[i], 1/size);
-        rootMeanSquare += Math.pow(distribution[i], 2);
-    }
-    var mean = mean.toFixed(5);
-    var meanDiff = distribution.map(el => parseFloat((el - mean).toFixed(5)));
-    return {
-        'mean': parseFloat(mean),
-        'geometricMean': parseFloat(geometricMean.toFixed(5)),
-        'rootMeanSquare': parseFloat(Math.pow(rootMeanSquare/size, 1/2).toFixed(5)),
-        'meanDiff': meanDiff
-    }
-}
-
-
-var ModeMedian = function(distribution) {
-    var size = distribution.length;
-    var sorted_distribution;
-    if (typeof distribution[0] === 'string') {
-        sorted_distribution = distribution.sort();
-    }
-    var sorted_distribution = distribution.sort((x, y) => x-y);
-
-    var mode = [sorted_distribution[0], 1];
-    var median;
-    var mid = sorted_distribution.length/2;
-    if ( sorted_distribution.length%2 != 0) {
-        median = [sorted_distribution[Math.floor(mid)]];
-    } else {
-        median = [sorted_distribution[mid-1], sorted_distribution[mid]];
-    }
-    var frequencies = [[(sorted_distribution[0]), 1]];
-    for (var i = 1; i < size; ++i ) {
-        if ( sorted_distribution[i] != frequencies[frequencies.length - 1][0]) {
-            frequencies.push([sorted_distribution[i], 1]);
-        } else {
-            frequencies[frequencies.length - 1][1] += 1;
-            mode = frequencies[frequencies.length -1][1] > mode[1] ? frequencies[frequencies.length -1]: mode;
-        }
-    }
-    if (frequencies.length/distribution.length > 0.85 && typeof distribution[0] === typeof 1) {
-        frequencies = 'It might not be appropriate to consider this random variable' +
-                      ' as categorical in your analysis: ' + (frequencies.length*100/distribution.length).toFixed(2) +
-                      '% values are unique.' ;
-    }
-    return {
-        'mode': {value: mode[0], frequency: mode[1]},
-        'median': {value: median},
-        'frequencies': frequencies,
-        'sorted': sorted_distribution,
-    }
-}
-
-
-module.exports = {
-    'Mean': Mean,
-    'Median': ModeMedian,
-}
-},{}]},{},[3]);
+//# sourceMappingURL=math.min.map
