+++ data_Insights_website/templates/includes/footer.html
@@ -1,4 +1,4 @@
-<div class="container-fluid footer">
+<div class="container-fluid" style='left: 0; bottom:0, position:fixed'>
     <div class="row" id="footer">
         <span class="text-sm" style='margin: 20px'>&copy;data.<span class="colored-text-red">Insights</span> 2018</span>
         <span class="text-sm ml-auto" style='margin: 20px'>GNU GPL v3.0</span>
