+++ data_Insights_website/templates/pages/app_load_success.html
@@ -63,6 +63,8 @@
             </div>
         </div>
         <div id="chart"></div>
+        
+    {% include 'includes/footer.html' %}
         </div>
     <script>
         var id = 0;
@@ -77,7 +79,6 @@
             charts.ecParams.ecParams.values = arg.values;
             charts.ecParams.ecParams.categories = arg.categories;
             //charts.ecParams.ecParams._data_generator = arg.data_generator;
-            console.log(arg.chartName);
             if(arg.chartName == 'line' || arg.chartName == 'area'){
                 var option = charts.line.chart(arg.xAxis, [arg.yAxis], ['line chart']);
             }else{
@@ -97,7 +98,6 @@
         }
         function create_chart(chartId){
             var chart_holder = document.createElement('div');
-            console.log("I am here adding");
             chart_holder.setAttribute('id', chartId);
             chart_holder.style.height = '500px';
             chart_holder.style.width = '1000px';
@@ -128,8 +128,6 @@
         };
         function filterCols(event){
             chart = document.querySelector('#chart_type').value;
-            console.log("chart name ");
-            console.log(chart);
             var _cols = {{cols|safe}}
             var _data = {{coldata|safe}}
             j = 0;
@@ -225,7 +223,6 @@
         data = {{ coldata|safe }}
         num_columns = {{ num_cols|safe }}
         columns = {{ cols|safe }}
-        console.log(num_columns)
         var mean_row= {'stats': 'Mean'};
         var geometricMean_row = {'stats': 'Geormetric Mean'};
         var rootMeanSquare_row = {'stats': 'Root Mean Square'};
@@ -325,5 +322,4 @@
             });
 
     </script>
-    {% include 'includes/footer.html' %}
 {% endblock %}
