+++ data_Insights_website/templates/pages/index.html
@@ -135,8 +135,8 @@
             "<span class='heading heading-lg'><span class='colored-text-red heading-xl'>A</span>NAL<span class='colored-text-curious-blue'>Y</span>SIS</span>",
             "<span class='heading heading-lg'><span class='colored-text-red'><span class='heading-xl'>I</span>NSIGHTS</span></span>"
         ],
-        typeSpeed: 70,
-        startDelay: 2000,
+        typeSpeed: 80,
+        startDelay: 3000,
         loop: true,
         bindInputFocusEvents: true
     }
