+++ data_Insights_website/templates/pages/viz.html
@@ -2,13 +2,13 @@
 <script src="https://cdnjs.cloudflare.com/ajax/libs/particles.js/2.0.0/particles.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/dropzone/5.4.0/min/dropzone.min.js"></script>
 <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/dropzone/5.4.0/dropzone.css" integrity="sha256-0Z6mOrdLEtgqvj7tidYQnCYWG3G2GAIpatAWKhDx+VM=" crossorigin="anonymous" /> {% endblock %} {% block content %} {% include 'includes/navigationbar.html' %} {% include 'includes/particles-bg.html' %}
-<div id="chart" style="height:500px; width:1200px"></div>
-<div id="chart0" style="height:500px; width:1200px"></div>
-<div id="chart1" style="height:500px; width:1200px"></div>
-<div id="chart2" style="height:500px; width:1200px"></div>
-<div id="chart3" style="height:500px; width:1200px"></div>
-<div id="chart4" style="height:500px; width:1200px"></div>
-<div id="chart5" style="height:500px; width:1200px"></div>
+<div id="chart" style="height:500px; width:1000px"></div>
+<div id="chart0" style="height:500px; width:1000px"></div>
+<div id="chart1" style="height:500px; width:1000px"></div>
+<div id="chart2" style="height:500px; width:1000px"></div>
+<div id="chart3" style="height:500px; width:1000px"></div>
+<div id="chart4" style="height:500px; width:1000px"></div>
+<div id="chart5" style="height:500px; width:1000px"></div>
 <script src="{% static 'js/lib/charts/echarts-en.min.js'%}"></script>
 <script src="{% static 'js/lib/charts/bundle.js'%}"></script>
 <script src="{% static 'js/quandl_test_data.js'%}"></script>
@@ -77,8 +77,8 @@
         myChart.setOption(option);
     }
 
-    option = charts.line.chart([-1.5, -0.75, 0, 0.75, 1.5, 2, 3.2], [
-        [-14.1014, -0.931596, 0, 0.931596, 14.1014, 10.152, 9.12]
+    option = charts.line.chart([-1.5, -0.75, 0, 0.75, 1.5, 2, 3.2, 4, 4.5, 5.67, 6, 6.8, 7.2, 8, 8.1, 9, 9.59], [
+        [-14.1014, -0.931596, 0, 0.931596, 14.1014, 10.152, 9.12, 5.987, 6.5, -3, 8, -12.56, 10, 11.5, 6, 11, 3.56]
     ], ['Interpolate'])
     var myChart3 = echarts.init(document.getElementById('chart1'), charts.mime.theme);
     myChart3.setOption(option);
